% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tempflow-basics.R
\name{add_points.tempflow}
\alias{add_points.tempflow}
\title{Is a function that takes a 'tempflow'-object and a vector of strings,
and adds new points, with the specified point-names, to the 'tempflow'-object.}
\usage{
\method{add_points}{tempflow}(tempFlow, newPointList, attrList = NULL,
  safe = TRUE)
}
\arguments{
\item{tempFlow}{A 'tempflow'-object.}

\item{attrList}{A vector to specify other vertex attributes. If left NULL, no attributes (other than name) will be assigned to the newly introduced points in time.
Otherwise, this list has to be structured as required by the function igraph::add.vertices.}

\item{safe}{If TRUE, it will be ensured that point-names remain unique; that point-names remain strings.
Additionally, the usual tempflow condition are checked (e.g. duplicate names, cycles, ... ).}

\item{addPointList}{A vector containing the strings that represent the names of the points that ought to be added.}
}
\value{
A new 'tempflow'-object, constructed by adding points with the specified names to the given 'tempflow'-object.
}
\description{
Is a function that takes a 'tempflow'-object and a vector of strings,
and adds new points, with the specified point-names, to the 'tempflow'-object.
}
