% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tempnetwork-tempflow-basics.R
\name{add_points.tempnetwork}
\alias{add_points.tempnetwork}
\title{Is a function that takes a 'tempnetwork'-object and a vector of strings,
and adds new points, with the specified point-names, to the 'tempflow'-object of the given 'tempnetwork'-object.
Moreover, the graphs corresponding to the newly added points will be embedded into the "graphList" of the 'tempnetwork'-object and
the equivalence relation will be exteded by by the equivalence relation given through the parameter "equivRel".}
\usage{
\method{add_points}{tempnetwork}(tempNetwork, newPointList, graphList,
  equivRel = NULL, attrList = NULL, safe = TRUE)
}
\arguments{
\item{tempNetwork}{A 'tempnetwork'-object.}

\item{graphList}{A named list containing 'igraph'-objects. The names of this list must correspond with the names of the newly added points.
Moreover, all graphs must have the same number of vertices.}

\item{equivRel}{A named list of named lists,
where the outer-level names must be identical to the outer-level names of the already existing equivalence-relation;
where the inner-level names must be identical with the names of the newly added points;
where the values must be identical to the vertex-names of the corresponding newly added graph.

Moreover, if the existing equivalence class is NULL and the parameter "equivRel" is NULL, then it must be the case that
the names of the vertices must be identical in each graph (including the already existing graphs).
Otherwise, the given equivalence relation will be merged with the already existing one.}

\item{attrList}{A vector to specify other vertex attributes. If left NULL, no attributes (other than name) will be assigned to the newly introduced points in time.
Otherwise, this list has to be structured as required by the function igraph::add.vertices.}

\item{safe}{If TRUE, it will be ensured that point-names remain unique; that point-names remain strings.
Additionally, the usual tempflow and tempnetwork condition are checked.}

\item{addPointList}{A vector containing the strings that represent the names of the points that ought to be added.}
}
\value{
A new 'tempnetwork'-object, containing a new 'tempflow'-object constructed by adding points with the specified names to the 'tempflow'-object of the given 'tempnetwork'-object;
containing a new "graphList" constructed by extending the existing "graphList" by the provided list of graphs;
containing a new "equivRel" constructed by extending the existing "equivRel" by the provided equivalence relation fragment.
}
\description{
Is a function that takes a 'tempnetwork'-object and a vector of strings,
and adds new points, with the specified point-names, to the 'tempflow'-object of the given 'tempnetwork'-object.
Moreover, the graphs corresponding to the newly added points will be embedded into the "graphList" of the 'tempnetwork'-object and
the equivalence relation will be exteded by by the equivalence relation given through the parameter "equivRel".
}
