% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tempnetwork-inputchecker.R
\name{base_check_equivalenceclass_is_exhaustive}
\alias{base_check_equivalenceclass_is_exhaustive}
\title{A function that checks,
if for each entry in the equivalence-class of the given equivalence-relation,
there exists a vertex with the same name in the corresponding graph from the given graph-list.}
\usage{
base_check_equivalenceclass_is_exhaustive(graphList, equivRel)
}
\arguments{
\item{graphList}{A named list of 'igraph'-objects.}

\item{equivRel}{A named list of named lists.}
}
\value{
The boolean value TRUE,
if for each entry in the equivalence-class of the given equivalence-relation,
there exists a vertex with the same name in the corresponding graph from the given graph-list.
}
\description{
A function that checks,
if for each entry in the equivalence-class of the given equivalence-relation,
there exists a vertex with the same name in the corresponding graph from the given graph-list.
}
