% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tempnetwork-inputchecker.R
\name{base_check_equivclass_disjoined}
\alias{base_check_equivclass_disjoined}
\title{A function that checks,
if the equivalence-classes in the given equivalence-relation are disjoined for each index (i.e. point).}
\usage{
base_check_equivclass_disjoined(equivRel)
}
\arguments{
\item{equivRel}{A named list of named lists.}
}
\value{
The boolean value TRUE,
if the equivalence-classes in the given equivalence-relation are disjoined for each index (i.e. point).
}
\description{
A function that checks,
if the equivalence-classes in the given equivalence-relation are disjoined for each index (i.e. point).
}
