% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tempnetwork-inputchecker.R
\name{base_check_equivclass_equal_length}
\alias{base_check_equivclass_equal_length}
\title{A function that checks,
if each equivalence-class in the given equivalence-relation has the same length.}
\usage{
base_check_equivclass_equal_length(equivRel)
}
\arguments{
\item{equivRel}{A named list of named lists.}
}
\value{
The boolean value TRUE,
if each equivalence-class in the given equivalence-relation has the same length.
}
\description{
A function that checks,
if each equivalence-class in the given equivalence-relation has the same length.
}
