% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tempnetwork-inputchecker.R
\name{base_check_equivclass_equal_length_to_graphs}
\alias{base_check_equivclass_equal_length_to_graphs}
\title{A function that checks,
if the number of vertices of each equvialence-class in the given equivalence-relation is identical to
the number of graphs in the given graph-list.}
\usage{
base_check_equivclass_equal_length_to_graphs(graphList, equivRel)
}
\arguments{
\item{graphList}{A named list of 'igraph'-objects.}

\item{equivRel}{A named list of named lists.}
}
\value{
The boolean value TRUE,
if the number of vertices of each equvialence-class in the given equivalence-relation is identical to
the number of graphs in the given graph-list.
}
\description{
A function that checks,
if the number of vertices of each equvialence-class in the given equivalence-relation is identical to
the number of graphs in the given graph-list.
}
