% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tempnetwork-inputchecker.R
\name{base_check_equivclass_index_is_unique}
\alias{base_check_equivclass_index_is_unique}
\title{A function that checks,
if the index set of each equivalence-class in the given equivalence-relation is unique.}
\usage{
base_check_equivclass_index_is_unique(equivRel)
}
\arguments{
\item{equivRel}{A named list of named lists.}
}
\value{
The boolean value TRUE,
if the index set of each equivalence-class in the given equivalence-relation is unique.
}
\description{
A function that checks,
if the index set of each equivalence-class in the given equivalence-relation is unique.
}
