% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tempnetwork-inputchecker.R
\name{base_check_equivrel_equal_length_to_vertices}
\alias{base_check_equivrel_equal_length_to_vertices}
\title{A function that checks,
if the number of vertices of each graph in the given graph-list is identical to
the number of equivalence-classes in the given equivalence-relation.}
\usage{
base_check_equivrel_equal_length_to_vertices(graphList, equivRel)
}
\arguments{
\item{graphList}{A named list of 'igraph'-objects.}

\item{equivRel}{A named list of named lists.}
}
\value{
The boolean value TRUE,
if the number of vertices of each graph in the given graph-list is identical to
the number of equivalence-classes in the given equivalence-relation.
}
\description{
A function that checks,
if the number of vertices of each graph in the given graph-list is identical to
the number of equivalence-classes in the given equivalence-relation.
}
