% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tempnetwork-inputchecker.R
\name{base_check_graphlist_equal_length}
\alias{base_check_graphlist_equal_length}
\title{A function that checks,
if the vertex set of each 'igraph'-object in the given graph-list has the same length.}
\usage{
base_check_graphlist_equal_length(graphList)
}
\arguments{
\item{graphList}{A named list of 'igraph'-objects.}
}
\value{
The boolean value TRUE,
if the vertex set of each 'igraph'-object in the given graph-list has the same length.
}
\description{
A function that checks,
if the vertex set of each 'igraph'-object in the given graph-list has the same length.
}
