% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tempnetwork-inputchecker.R
\name{base_check_graphlist_equivclass_index_equality}
\alias{base_check_graphlist_equivclass_index_equality}
\title{A function that checks,
if the indices of each equivalence-class from the given equivalence-relation,
are identical to the strings in the given string-vector.}
\usage{
base_check_graphlist_equivclass_index_equality(graphListNameList, equivRel)
}
\arguments{
\item{graphListNameList}{A vector of strings.}

\item{equivRel}{A named list of named lists.}
}
\value{
The boolean value TRUE,
if the indices of each equivalence-class from the given equivalence-relation,
are identical to the strings in the given string-vector.
}
\description{
A function that checks,
if the indices of each equivalence-class from the given equivalence-relation,
are identical to the strings in the given string-vector.
}
