% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tempnetwork-inputchecker.R
\name{base_check_graphlist_index_is_unique}
\alias{base_check_graphlist_index_is_unique}
\title{A function that checks,
if the given names are unique.}
\usage{
base_check_graphlist_index_is_unique(graphListNameList)
}
\arguments{
\item{graphListNameList}{A vector of strings.}
}
\value{
The boolean value TRUE,
if the given names are unique.
}
\description{
A function that checks,
if the given names are unique.
}
