% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tempnetwork-inputchecker.R
\name{base_check_graphlist_not_empty}
\alias{base_check_graphlist_not_empty}
\title{A function that checks, if the given graph-list contains at least one entry.}
\usage{
base_check_graphlist_not_empty(graphList)
}
\arguments{
\item{graphList}{A named list of 'igraph'-objects.}
}
\value{
The boolean value TRUE, if the given graph-list contains at least one entry.
}
\description{
A function that checks, if the given graph-list contains at least one entry.
}
