% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tempnetwork-inputchecker.R
\name{base_check_graphlist_vertex_equality}
\alias{base_check_graphlist_vertex_equality}
\title{A function that checks,
if the vertex sets of all graphs in the givern graph-list are identical.}
\usage{
base_check_graphlist_vertex_equality(graphList)
}
\arguments{
\item{graphList}{A named list of 'igraph'-objects.}
}
\value{
The boolean value TRUE,
if the vertex sets of all graphs in the givern graph-list are identical.
}
\description{
A function that checks,
if the vertex sets of all graphs in the givern graph-list are identical.
}
