% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tempnetwork-inputchecker.R
\name{base_check_graphlist_vertex_unique}
\alias{base_check_graphlist_vertex_unique}
\title{A function that checks,
if the names of each vertex in each graph of the given graph-list is unique for its graph.}
\usage{
base_check_graphlist_vertex_unique(graphList)
}
\arguments{
\item{graphList}{A named list of 'igraph'-objects.}
}
\value{
The boolean value TRUE,
if the names of each vertex in each graph of the given graph-list is unique for its graph.
}
\description{
A function that checks,
if the names of each vertex in each graph of the given graph-list is unique for its graph.
}
