% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tempnetwork-inputchecker.R
\name{base_check_graphs_have_names}
\alias{base_check_graphs_have_names}
\title{A function that checks, if the 'igraphs'-object in the given graph-list have the vertex attribute "name".}
\usage{
base_check_graphs_have_names(graphList)
}
\arguments{
\item{graphList}{A named list of 'igraph'-objects.}
}
\value{
The boolean value TRUE, if the 'igraphs'-object in the given graph-list have the vertex attribute "name".
}
\description{
A function that checks, if the 'igraphs'-object in the given graph-list have the vertex attribute "name".
}
