% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tempnetwork-inputchecker.R
\name{base_check_is_pointsubset}
\alias{base_check_is_pointsubset}
\title{A function that checks,
if the points in the given 'tempflow'-object "subTempFlow" are a subset of the points in  the  given 'tempflow'-object "supTempFlow".}
\usage{
base_check_is_pointsubset(supTempFlow, subTempFlow)
}
\arguments{
\item{supTempFlow}{A 'tempnetwork'-object.}

\item{subTempFlow}{A 'tempflow'-object.}
}
\value{
The boolean value TRUE,
if the points in the given 'tempflow'-object "subTempFlow" are a subset of the points in  the  given 'tempflow'-object "supTempFlow"
}
\description{
A function that checks,
if the points in the given 'tempflow'-object "subTempFlow" are a subset of the points in  the  given 'tempflow'-object "supTempFlow".
}
