% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tempnetwork-inputchecker.R
\name{base_check_is_subtempflow}
\alias{base_check_is_subtempflow}
\title{A function that checks,
if the given 'tempflow'-object "subTempFlow" is a subtempflow the  given 'tempflow'-object "supTempFlow".
(That is, the underlying graph of the prior is a subgraph of the underlying graph of the latter)}
\usage{
base_check_is_subtempflow(supTempFlow, subTempFlow)
}
\arguments{
\item{supTempFlow}{A 'tempnetwork'-object.}

\item{subTempFlow}{A 'tempflow'-object.}
}
\value{
The boolean value TRUE,
if the given 'tempflow'-object "subTempFlow" is a subtempflow the  given 'tempflow'-object "supTempFlow".
}
\description{
A function that checks,
if the given 'tempflow'-object "subTempFlow" is a subtempflow the  given 'tempflow'-object "supTempFlow".
(That is, the underlying graph of the prior is a subgraph of the underlying graph of the latter)
}
