% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tempnetwork-inputchecker.R
\name{base_check_tempflow_graphlist_index_equality}
\alias{base_check_tempflow_graphlist_index_equality}
\title{A function that checks,
if the names in one given string-vector
are identical to the names in the other given string-vector.}
\usage{
base_check_tempflow_graphlist_index_equality(tempFlowPointNameList,
  graphListNameList)
}
\arguments{
\item{tempFlowPointNameList}{A vector of strings.}

\item{graphListNameList}{A vector of strings.}

\item{equivRel}{A named list of named lists.}
}
\value{
The boolean value TRUE,
if the names in one given string-vector
are identical to the names in the other given string-vector.
}
\description{
A function that checks,
if the names in one given string-vector
are identical to the names in the other given string-vector.
}
