% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tempnetwork-inputchecker.R
\name{base_check_tempflow_not_empty}
\alias{base_check_tempflow_not_empty}
\title{A function that checks, if the given 'tempflow'-object contains at least one point.}
\usage{
base_check_tempflow_not_empty(tempFlow)
}
\arguments{
\item{tempFlow}{A 'tempflow'-object.}
}
\value{
The boolean value TRUE, if the given 'tempflow'-object contains at least one point.
}
\description{
A function that checks, if the given 'tempflow'-object contains at least one point.
}
