% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tempnetwork-getter.R
\name{base_get_tempgraph_vertex_name}
\alias{base_get_tempgraph_vertex_name}
\title{Concatenate the strings tVertexName, pointName and vertexName with an empty separator.}
\usage{
base_get_tempgraph_vertex_name(tVertexName, pointName, vertexName)
}
\arguments{
\item{tVertexName}{A string.}

\item{pointName}{A string.}

\item{vertexName}{A string.}

\item{tempNetwork}{A 'tempnetwork'-object.}
}
\value{
A concatenated string composed of the strings tVertexName, pointName and vertexName joined through the empty seperator.
}
\description{
Concatenate the strings tVertexName, pointName and vertexName with an empty separator.
}
