% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tempflow-predicates.R
\name{between.tempflow}
\alias{between.tempflow}
\title{Is a function that takes a 'tempflow'-object and three points and tests whether "checkPoint" is strictly between "startPoint" and end"Point,
given the order implied by the 'tempflow'-object represented by the 'tempflow'-object.}
\usage{
\method{between}{tempflow}(tempFlow, startPoint, endPoint, checkPoint,
  safe = TRUE)
}
\arguments{
\item{tempFlow}{A 'tempflow'-object.}

\item{startPoint}{A point-name (or point-id).}

\item{endPoint}{A point-name (or point-id).}

\item{checkPoint}{A point-name (or point-id).}

\item{safe}{If TRUE, it will be tested, whether the given point-names (or point-ids) are present in the 'tempflow'-object.}
}
\value{
Is TRUE if "checkPoint" is strictly between "startPoint" and "endPoint", else FALSE will be returned.
}
\description{
Is a function that takes a 'tempflow'-object and three points and tests whether "checkPoint" is strictly between "startPoint" and end"Point,
given the order implied by the 'tempflow'-object represented by the 'tempflow'-object.
}
