% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tempnetwork-tempflow-predicates.R
\name{between.tempnetwork}
\alias{between.tempnetwork}
\title{Is a function that takes a 'tempnetwork'-object and three points and tests whether "checkPoint" is strictly between "startPoint" and end"Point,
given the order induced by the 'tempflow'-object of the given 'tempnetwork'-object.}
\usage{
\method{between}{tempnetwork}(tempNetwork, startPoint, endPoint,
  checkPoint, safe = TRUE)
}
\arguments{
\item{tempNetwork}{A 'tempnetwork'-object.}

\item{startPoint}{A point-name (or point-id).}

\item{endPoint}{A point-name (or point-id).}

\item{checkPoint}{A point-name (or point-id).}

\item{safe}{If TRUE, it will be tested, whether the given point-names (or point-ids) are present in the 'tempflow'-object of the given 'tempnetwork'-object.}
}
\value{
Is TRUE if "checkPoint" is strictly between "startPoint" and "endPoint", else FALSE will be returned.
}
\description{
Is a function that takes a 'tempnetwork'-object and three points and tests whether "checkPoint" is strictly between "startPoint" and end"Point,
given the order induced by the 'tempflow'-object of the given 'tempnetwork'-object.
}
