% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tempflow-inputchecker.R
\name{check_are_names_strings.tempflow}
\alias{check_are_names_strings.tempflow}
\title{A function that checks, if the value of the 'name'-attribute of each point in the 'tempflow'-object is of type 'string'.}
\usage{
\method{check_are_names_strings}{tempflow}(tempFlow)
}
\arguments{
\item{tempFlow}{A 'tempflow'-object.}
}
\value{
The boolean value TRUE, if the value of the 'name'-attribute of each point in the 'tempflow'-object is of type 'string'
}
\description{
A function that checks, if the value of the 'name'-attribute of each point in the 'tempflow'-object is of type 'string'.
}
