% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tempflow-inputchecker.R
\name{check_are_names_unique.tempflow}
\alias{check_are_names_unique.tempflow}
\title{A function that checks, if there are two points in 'tempflow'-object, that are assigned the same name.}
\usage{
\method{check_are_names_unique}{tempflow}(tempFlow)
}
\arguments{
\item{tempFlow}{A 'tempflow'-object.}
}
\value{
returns TRUE, if every point in the tempflow object has a unique 'name'-attribute.
}
\description{
A function that checks, if there are two points in 'tempflow'-object, that are assigned the same name.
}
