% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tempflow-inputchecker.R
\name{check_are_weights_nonnegative.tempflow}
\alias{check_are_weights_nonnegative.tempflow}
\title{A function that checks, if the value of the 'weight'-attribute of each point has a non-negative value.}
\usage{
\method{check_are_weights_nonnegative}{tempflow}(tempFlow)
}
\arguments{
\item{tempFlow}{A 'tempflow'-object.}
}
\value{
The boolean value TRUE, if every point in the 'tempflow'-object has a non-negative 'weight'-attribute.
}
\description{
A function that checks, if the value of the 'weight'-attribute of each point has a non-negative value.
}
