% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tempflow-inputchecker.R
\name{check_are_weights_numeric.tempflow}
\alias{check_are_weights_numeric.tempflow}
\title{A function that checks, if the value of the 'weight'-attribute of each point is a numerical value.}
\usage{
\method{check_are_weights_numeric}{tempflow}(tempFlow)
}
\arguments{
\item{tempFlow}{A 'tempflow'-object.}
}
\value{
The boolean value TRUE, if every point in the 'tempflow'-object has a numeric 'weight'-attribute (excluding NA, NaN and NULL values).
}
\description{
A function that checks, if the value of the 'weight'-attribute of each point is a numerical value.
}
