% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tempnetwork-inputchecker.R
\name{check_edge_exists.tempnetwork}
\alias{check_edge_exists.tempnetwork}
\title{The corresponding entries in "pointNameList", "startTVertexNameList" and "endTVertexNameList", form a (graph,start-vertex,end-vertex)-pair.
It will be checked that there exists an edge in between the start- and end-vertex in the specified graph from the graph-list of the given 'tempnetwork'-object.}
\usage{
\method{check_edge_exists}{tempnetwork}(tempNetwork, pointNameList,
  startTVertexNameList, endTVertexNameList)
}
\arguments{
\item{tempNetwork}{A 'tempnetwork'-object.}

\item{pointNameList}{A vector of strings, representing the names of graphs in the graph-list of the given 'tempnetwork'-object.}

\item{startTVertexNameList}{A vector of strings, representing the names of vertices in graphs from the graph-list of the given 'tempnetwork'-object. It has same length as the vector "pointNameList".}

\item{endTVertexNameList}{A vector of strings, representing the names of vertices in graphs from the graph-list of the given 'tempnetwork'-object. It has same length as the vector "pointNameList".}
}
\value{
The boolean value TRUE,
if between every specified (graph,start-vertex,end-vertex)-pair, there exists an edge in between the start- and end-vertex in the specified graph from the graph-list of the given 'tempnetwork'-object.
}
\description{
The corresponding entries in "pointNameList", "startTVertexNameList" and "endTVertexNameList", form a (graph,start-vertex,end-vertex)-pair.
It will be checked that there exists an edge in between the start- and end-vertex in the specified graph from the graph-list of the given 'tempnetwork'-object.
}
