% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tempnetwork-inputchecker.R
\name{check_equivalenceclass_is_exhaustive.tempnetwork}
\alias{check_equivalenceclass_is_exhaustive.tempnetwork}
\title{A function that checks,
if for each entry in the equivalence-class of the equivalence-relation of the given 'tempnetwork'-object,
 there exists a vertex with the same name in the corresponding graph from the graph-list of the given 'tempnetwork'-object.}
\usage{
\method{check_equivalenceclass_is_exhaustive}{tempnetwork}(tempNetwork)
}
\arguments{
\item{tempNetwork}{A 'tempnetwork'-object.}
}
\value{
The boolean value TRUE,
if for each entry in the equivalence-class of the equivalence-relation of the given 'tempnetwork'-object,
 there exists a vertex with the same name in the corresponding graph from the graph-list of the given 'tempnetwork'-object.
}
\description{
A function that checks,
if for each entry in the equivalence-class of the equivalence-relation of the given 'tempnetwork'-object,
 there exists a vertex with the same name in the corresponding graph from the graph-list of the given 'tempnetwork'-object.
}
