% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tempnetwork-inputchecker.R
\name{check_equivclass_disjoined.tempnetwork}
\alias{check_equivclass_disjoined.tempnetwork}
\title{A function that checks,
if the equivalence-classes in the equivalence-relation of the given 'tempnetwork'-object are disjoined for each index (i.e. point).}
\usage{
\method{check_equivclass_disjoined}{tempnetwork}(tempNetwork)
}
\arguments{
\item{tempNetwork}{A 'tempnetwork'-object.}
}
\value{
The boolean value TRUE,
if the equivalence-classes in the equivalence-relation of the given 'tempnetwork'-object are disjoined for each index (i.e. point).
}
\description{
A function that checks,
if the equivalence-classes in the equivalence-relation of the given 'tempnetwork'-object are disjoined for each index (i.e. point).
}
