% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tempnetwork-inputchecker.R
\name{check_equivclass_equal_length.tempnetwork}
\alias{check_equivclass_equal_length.tempnetwork}
\title{A function that checks,
if each equivalence-class in the equivalence-relation of the given 'tempnetwork'-object has the same length.}
\usage{
check_equivclass_equal_length.tempnetwork(tempNetwork)
}
\arguments{
\item{tempNetwork}{A 'tempnetwork'-object.}
}
\value{
The boolean value TRUE,
if each equivalence-class in the equivalence-relation of the given 'tempnetwork'-object has the same length.
}
\description{
A function that checks,
if each equivalence-class in the equivalence-relation of the given 'tempnetwork'-object has the same length.
}
