% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tempnetwork-inputchecker.R
\name{check_equivclass_equal_length_to_graphs.tempnetwork}
\alias{check_equivclass_equal_length_to_graphs.tempnetwork}
\title{A function that checks,
if the number of vertices of each equvialence-class in the equivalence-relation of the given 'tempnetwork'-object is identical to
the number of graphs in the graph-list of the given 'tempnetwork'-object.}
\usage{
\method{check_equivclass_equal_length_to_graphs}{tempnetwork}(tempNetwork)
}
\arguments{
\item{tempNetwork}{A 'tempnetwork'-object.}
}
\value{
The boolean value TRUE,
if the number of vertices of each equvialence-class in the equivalence-relation of the given 'tempnetwork'-object is identical to
the number of graphs in the graph-list of the given 'tempnetwork'-object.
}
\description{
A function that checks,
if the number of vertices of each equvialence-class in the equivalence-relation of the given 'tempnetwork'-object is identical to
the number of graphs in the graph-list of the given 'tempnetwork'-object.
}
