% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tempnetwork-inputchecker.R
\name{check_equivclass_index_is_unique.tempnetwork}
\alias{check_equivclass_index_is_unique.tempnetwork}
\title{A function that checks,
if the index set of each equivalence-class in the equivalence-relation of the given 'tempnetwork'-object is unique.}
\usage{
\method{check_equivclass_index_is_unique}{tempnetwork}(tempNetwork)
}
\arguments{
\item{tempNetwork}{A 'tempnetwork'-object.}
}
\value{
The boolean value TRUE,
if the index set of each equivalence-class in the equivalence-relation of the given 'tempnetwork'-object is unique.
}
\description{
A function that checks,
if the index set of each equivalence-class in the equivalence-relation of the given 'tempnetwork'-object is unique.
}
