% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tempnetwork-inputchecker.R
\name{check_equivrel_index_is_unique.tempnetwork}
\alias{check_equivrel_index_is_unique.tempnetwork}
\title{A function that checks,
if each equivalence-class in the equivalence-relation of the given 'tempnetwork'-object has a unique name.}
\usage{
\method{check_equivrel_index_is_unique}{tempnetwork}(tempNetwork)
}
\arguments{
\item{tempNetwork}{A 'tempnetwork'-object.}
}
\value{
The boolean value TRUE,
if each equivalence-class in the equivalence-relation of the given 'tempnetwork'-object has a unique name.
}
\description{
A function that checks,
if each equivalence-class in the equivalence-relation of the given 'tempnetwork'-object has a unique name.
}
