% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tempnetwork-inputchecker.R
\name{check_graphlist_equal_length.tempnetwork}
\alias{check_graphlist_equal_length.tempnetwork}
\title{A function that checks,
if the vertex set of each 'igraph'-object in the graph-list of the given 'tempnetwork'-object has the same length.}
\usage{
\method{check_graphlist_equal_length}{tempnetwork}(tempNetwork)
}
\arguments{
\item{tempNetwork}{A 'tempnetwork'-object.}
}
\value{
The boolean value TRUE,
if the vertex set of each 'igraph'-object in the graph-list of the given 'tempnetwork'-object has the same length.
}
\description{
A function that checks,
if the vertex set of each 'igraph'-object in the graph-list of the given 'tempnetwork'-object has the same length.
}
