% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tempnetwork-inputchecker.R
\name{check_graphlist_equivclass_index_equality.tempnetwork}
\alias{check_graphlist_equivclass_index_equality.tempnetwork}
\title{A function that checks,
if the indices of each equivalence-class from the equivalence-relation of the given 'tempnetwork'-object,
are identical to the indices from the graph-list of the given 'tempnetwork'-object.}
\usage{

  \method{check_graphlist_equivclass_index_equality}{tempnetwork}(tempNetwork)
}
\arguments{
\item{tempNetwork}{A 'tempnetwork'-object.}
}
\value{
The boolean value TRUE,
if the indices of each equivalence-class from the equivalence-relation of the given 'tempnetwork'-object,
are identical to the indices from the graph-list of the given 'tempnetwork'-object.
}
\description{
A function that checks,
if the indices of each equivalence-class from the equivalence-relation of the given 'tempnetwork'-object,
are identical to the indices from the graph-list of the given 'tempnetwork'-object.
}
