% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tempnetwork-inputchecker.R
\name{check_graphlist_index_is_unique.tempnetwork}
\alias{check_graphlist_index_is_unique.tempnetwork}
\title{A function that checks,
if the names of the graphs in the graph-list of the given 'tempnetwork'-object are unique.}
\usage{
\method{check_graphlist_index_is_unique}{tempnetwork}(tempNetwork)
}
\arguments{
\item{tempNetwork}{A 'tempnetwork'-object.}
}
\value{
The boolean value TRUE,
if the names of the graphs in the graph-list of the given 'tempnetwork'-object are unique.
}
\description{
A function that checks,
if the names of the graphs in the graph-list of the given 'tempnetwork'-object are unique.
}
