% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tempnetwork-inputchecker.R
\name{check_graphlist_not_empty.tempnetwork}
\alias{check_graphlist_not_empty.tempnetwork}
\title{A function that checks, if the graph-list of the given 'tempnetwork'-object contains at least one entry.}
\usage{
\method{check_graphlist_not_empty}{tempnetwork}(tempNetwork)
}
\arguments{
\item{tempNetwork}{A 'tempnetwork'-object.}
}
\value{
The boolean value TRUE, if the graph-list of the given 'tempnetwork'-object contains at least one entry.
}
\description{
A function that checks, if the graph-list of the given 'tempnetwork'-object contains at least one entry.
}
