% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tempnetwork-inputchecker.R
\name{check_graphlist_vertex_equality.tempnetwork}
\alias{check_graphlist_vertex_equality.tempnetwork}
\title{A function that checks,
if the vertex sets of all graphs in the graph-list of the given 'tempnetwork'-object are identical.}
\usage{
\method{check_graphlist_vertex_equality}{tempnetwork}(tempNetwork)
}
\arguments{
\item{tempNetwork}{A 'tempnetwork'-object.}
}
\value{
The boolean value TRUE,
if the vertex sets of all graphs in the graph-list of the given 'tempnetwork'-object are identical.
}
\description{
A function that checks,
if the vertex sets of all graphs in the graph-list of the given 'tempnetwork'-object are identical.
}
