% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tempnetwork-inputchecker.R
\name{check_graphlist_vertex_sets.tempnetwork}
\alias{check_graphlist_vertex_sets.tempnetwork}
\title{If the equivalence-relation of the given 'tempnetwork'-object is NULL,
this function checks, if the vertex sets of all graphs in the graph-list of the given 'tempnetwork'-object are identical.
Otherwise, it checks if the vertex set of each 'igraph'-object in the graph-list of the given 'tempnetwork'-object has the same length.}
\usage{
\method{check_graphlist_vertex_sets}{tempnetwork}(tempNetwork,
  throwError = TRUE)
}
\arguments{
\item{tempNetwork}{A 'tempnetwork'-object.}

\item{throwError}{If TRUE the respective errors will be thrown. Otherwise, a truth value will be returned.}
}
\value{
If the equivalence-relation of the given 'tempnetwork'-object is NULL,
this function returns TRUE, if the vertex sets of all graphs in the graph-list of the given 'tempnetwork'-object are identical.
Otherwise, it returns TRUE if the vertex set of each 'igraph'-object in the graph-list of the given 'tempnetwork'-object has the same length.
}
\description{
If the equivalence-relation of the given 'tempnetwork'-object is NULL,
this function checks, if the vertex sets of all graphs in the graph-list of the given 'tempnetwork'-object are identical.
Otherwise, it checks if the vertex set of each 'igraph'-object in the graph-list of the given 'tempnetwork'-object has the same length.
}
