% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tempnetwork-inputchecker.R
\name{check_graphlist_vertex_unique.tempnetwork}
\alias{check_graphlist_vertex_unique.tempnetwork}
\title{A function that checks,
if the names of each vertex in each graph of the graph-list of the given 'tempnetwork'-object is unique for its graph.}
\usage{
\method{check_graphlist_vertex_unique}{tempnetwork}(tempNetwork)
}
\arguments{
\item{tempNetwork}{A 'tempnetwork'-object.}
}
\value{
The boolean value TRUE,
if the names of each vertex in each graph of the graph-list of the given 'tempnetwork'-object is unique for its graph.
}
\description{
A function that checks,
if the names of each vertex in each graph of the graph-list of the given 'tempnetwork'-object is unique for its graph.
}
