% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tempnetwork-inputchecker.R
\name{check_graphs_have_names.tempnetwork}
\alias{check_graphs_have_names.tempnetwork}
\title{A function that checks,
if the 'igraphs'-object in the graph-list of the given 'tempnetwork'-object have the vertex attribute "name".}
\usage{
\method{check_graphs_have_names}{tempnetwork}(tempNetwork)
}
\arguments{
\item{tempNetwork}{A 'tempnetwork'-object.}
}
\value{
The boolean value TRUE, if the 'igraphs'-object in the graph-list of the given 'tempnetwork'-object have the vertex attribute "name".
}
\description{
A function that checks,
if the 'igraphs'-object in the graph-list of the given 'tempnetwork'-object have the vertex attribute "name".
}
