% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tempflow-inputchecker.R
\name{check_has_loop.tempflow}
\alias{check_has_loop.tempflow}
\title{A function that checks, if the given point (specified by either name or id) has a loop.}
\usage{
\method{check_has_loop}{tempflow}(tempFlow, inpPoint, safe = TRUE)
}
\arguments{
\item{tempFlow}{A 'tempflow'-object.}

\item{inpPoint}{An integer or string value.}

\item{safe}{if TRUE, the existence of the point will be ensured.}
}
\value{
The boolean value TRUE, if the given point (specified by either name or id) has a loop.
}
\description{
A function that checks, if the given point (specified by either name or id) has a loop.
}
