% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tempflow-inputchecker.R
\name{check_is_igraph_object.tempflow}
\alias{check_is_igraph_object.tempflow}
\title{A function that checks, if the object assigned to the 'tempflow'-field "graph" is an 'igraph'-object.}
\usage{
\method{check_is_igraph_object}{tempflow}(tempFlow)
}
\arguments{
\item{tempFlow}{A 'tempflow'-object.}
}
\value{
The boolean value TRUE, if the object assigned to the 'tempflow'-field "graph" is an 'igraph'-object.
}
\description{
A function that checks, if the object assigned to the 'tempflow'-field "graph" is an 'igraph'-object.
}
