% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tempnetwork-inputchecker.R
\name{check_is_pointsubset.tempnetwork}
\alias{check_is_pointsubset.tempnetwork}
\title{A function that checks,
if the points in the given 'tempflow'-object are a subset of the points in the 'tempflow'-object from the given 'tempnetwork'-object.}
\usage{
\method{check_is_pointsubset}{tempnetwork}(tempNetwork, tempFlow)
}
\arguments{
\item{tempNetwork}{A 'tempnetwork'-object.}

\item{tempFlow}{A 'tempflow'-object.}
}
\value{
The boolean value TRUE,
if the points in the given 'tempflow'-object are a subset of the points in the 'tempflow'-object from the given 'tempnetwork'-object.
}
\description{
A function that checks,
if the points in the given 'tempflow'-object are a subset of the points in the 'tempflow'-object from the given 'tempnetwork'-object.
}
