% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tempflow-inputchecker.R
\name{check_is_simple.tempflow}
\alias{check_is_simple.tempflow}
\title{A function that checks, if the 'igraph'-object assigned to the 'tempflow'-field "graph" is a simple.}
\usage{
\method{check_is_simple}{tempflow}(tempFlow, ignoreLoops = TRUE)
}
\arguments{
\item{tempFlow}{A 'tempflow'-object.}

\item{ignoreLoops}{If TRUE, the graph will be considered simple, even if it contains loops. That is, only duplicate edges will be detected.}
}
\value{
The boolean value TRUE, if the 'igraph'-object assigned to the 'tempflow'-field "graph" is a simple.
}
\description{
A function that checks, if the 'igraph'-object assigned to the 'tempflow'-field "graph" is a simple.
}
