% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tempnetwork-inputchecker.R
\name{check_is_subtempflow.tempnetwork}
\alias{check_is_subtempflow.tempnetwork}
\title{A function that checks,
if the given 'tempflow'-object is a subtempflow the 'tempflow'-object from the given 'tempnetwork'-object.
(That is, the underlying graph of the prior is a subgraph of the underlying graph of the latter)}
\usage{
\method{check_is_subtempflow}{tempnetwork}(tempNetwork, tempFlow)
}
\arguments{
\item{tempNetwork}{A 'tempnetwork'-object.}

\item{tempFlow}{A 'tempflow'-object.}
}
\value{
The boolean value TRUE,
if the given 'tempflow'-object is a subtempflow the 'tempflow'-object from the given 'tempnetwork'-object.
}
\description{
A function that checks,
if the given 'tempflow'-object is a subtempflow the 'tempflow'-object from the given 'tempnetwork'-object.
(That is, the underlying graph of the prior is a subgraph of the underlying graph of the latter)
}
