% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tempflow-inputchecker.R
\name{check_names_for_tempflow.tempflow}
\alias{check_names_for_tempflow.tempflow}
\title{A function that checks, if the 'igraph'-object assigned to the tempflow-field 'graph', has unique vertex names (of type string).}
\usage{
\method{check_names_for_tempflow}{tempflow}(tempFlow, throwError = TRUE)
}
\arguments{
\item{tempFlow}{A 'tempflow'-object.}

\item{throwError}{If TRUE, violating the criteria above will result in the throwing of an error. Otherwise, FALSE will be returned.}
}
\value{
The boolean value TRUE, if the criteria above are satisfied. Otherwise, FALSE or an exception will be returned.
}
\description{
A function that checks, if the 'igraph'-object assigned to the tempflow-field 'graph', has unique vertex names (of type string).
}
