% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tempnetwork-inputchecker.R
\name{check_null_values.tempnetwork}
\alias{check_null_values.tempnetwork}
\title{A function that checks,
if the 'tempflow'-object of the given 'tempnetwork'-object contains at least one point and
the graph-list of the given 'tempnetwork'-object contains at least one entry.}
\usage{
\method{check_null_values}{tempnetwork}(tempNetwork, throwError = TRUE)
}
\arguments{
\item{tempNetwork}{A 'tempnetwork'-object.}

\item{throwError}{If TRUE the respective errors will be thrown. Otherwise, a truth value will be returned.}
}
\value{
The boolean value TRUE,
if the 'tempflow'-object of the given 'tempnetwork'-object contains at least one point and
the graph-list of the given 'tempnetwork'-object contains at least one entry.
}
\description{
A function that checks,
if the 'tempflow'-object of the given 'tempnetwork'-object contains at least one point and
the graph-list of the given 'tempnetwork'-object contains at least one entry.
}
