% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tempflow-inputchecker.R
\name{check_point_id.tempflow}
\alias{check_point_id.tempflow}
\title{A function that checks, if the given point-id corresponds to a vertex-id in the underlying graph.}
\usage{
\method{check_point_id}{tempflow}(tempFlow, pointID)
}
\arguments{
\item{tempFlow}{A 'tempflow'-object.}

\item{pointID}{An integer value.}
}
\value{
The boolean value TRUE, if the given integer corresponds to a vertex-id in the underlying 'igraph'-object.
}
\description{
A function that checks, if the given point-id corresponds to a vertex-id in the underlying graph.
}
