% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tempflow-inputchecker.R
\name{check_point_names.tempflow}
\alias{check_point_names.tempflow}
\title{A function that checks, if the given point-name occurs as a name-attribute of some vertex in the underlying graph.}
\usage{
\method{check_point_names}{tempflow}(tempFlow, pointNameList)
}
\arguments{
\item{tempFlow}{A 'tempflow'-object.}

\item{pointID}{A string value.}
}
\value{
The boolean value TRUE, if the given string occurs as a name-attribute of some vertex in the underlying 'igraph'-object.
}
\description{
A function that checks, if the given point-name occurs as a name-attribute of some vertex in the underlying graph.
}
