% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tempnetwork-inputchecker.R
\name{check_points_and_tvertices.tempnetwork}
\alias{check_points_and_tvertices.tempnetwork}
\title{This function simply combines the function "check_tvertices_exists" and "check_points_exists".}
\usage{
\method{check_points_and_tvertices}{tempnetwork}(tempNetwork,
  pointNameList = NULL, tVertexNameList = NULL, throwError = TRUE)
}
\arguments{
\item{tempNetwork}{A 'tempnetwork'-object.}

\item{pointNameList}{A vector of strings representing the names of points in the 'tempflow'-object of the given 'tempnetwork'-object. If NULL this input will be ignored.}

\item{tVertexNameList}{A vector of strings representing the names of equivalence-classes in the equivalence-relation of the given 'tempnetwork'-object. If NULL this input will be ignored.}

\item{throwError}{If TRUE the respective errors will be thrown. Otherwise, a truth value will be returned.}
}
\value{
The boolean value TRUE,
if the equivalence-classes specified in "tVertexNameList" exist in the equivalence-relation of the given 'tempnetwork'-object.
}
\description{
This function simply combines the function "check_tvertices_exists" and "check_points_exists".
}
