% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tempflow-inputchecker.R
\name{check_step_id.tempflow}
\alias{check_step_id.tempflow}
\title{A function that checks, if the given step-id corresponds to an edge-id in the underlying graph.}
\usage{
\method{check_step_id}{tempflow}(tempFlow, stepID)
}
\arguments{
\item{tempFlow}{A 'tempflow'-object.}

\item{pointID}{An integer value.}
}
\value{
The boolean value TRUE, if the given integer corresponds to an edge-id in the underlying 'igraph'-object.
}
\description{
A function that checks, if the given step-id corresponds to an edge-id in the underlying graph.
}
