% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tempflow-inputchecker.R
\name{check_step_ids.tempflow}
\alias{check_step_ids.tempflow}
\title{A function that checks, if each step-id in the given vector corresponds to an edge-id in the underlying graph.}
\usage{
\method{check_step_ids}{tempflow}(tempFlow, stepIDList)
}
\arguments{
\item{tempFlow}{A 'tempflow'-object.}

\item{pointID}{A vector of integer values.}
}
\value{
The boolean value TRUE, if each integer in the given vector corresponds to an edge-id in the underlying 'igraph'-object.
}
\description{
A function that checks, if each step-id in the given vector corresponds to an edge-id in the underlying graph.
}
