% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tempnetwork-inputchecker.R
\name{check_tempflow}
\alias{check_tempflow}
\title{A function that checks, if the given object is a 'tempflow'-object.}
\usage{
check_tempflow(tempNetwork)
}
\arguments{
\item{tempFlow}{A 'tempflow'-object.}
}
\value{
The boolean value TRUE, if the given object is a 'tempflow'-object
}
\description{
A function that checks, if the given object is a 'tempflow'-object.
}
