% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tempnetwork-inputchecker.R
\name{check_tempflow_graphlist_equivrel_correspondence.tempnetwork}
\alias{check_tempflow_graphlist_equivrel_correspondence.tempnetwork}
\title{A function that checks,
that the names of the graphs in the graph-list of the given 'tempnetwork'-object
are identical to names of the points from the 'tempflow'-object of the given 'tempnetwork'-object.
and that the indices of each equivalence-class from the equivalence-relation of the given 'tempnetwork'-object,
are identical to names of the points from the 'tempflow'-object of the given 'tempnetwork'-object.}
\usage{

  \method{check_tempflow_graphlist_equivrel_correspondence}{tempnetwork}(tempNetwork,
  throwError = TRUE)
}
\arguments{
\item{tempNetwork}{A 'tempnetwork'-object.}

\item{throwError}{If TRUE the respective errors will be thrown. Otherwise, a truth value will be returned.}
}
\value{
The boolean value TRUE,
if the names of the graphs in the graph-list of the given 'tempnetwork'-object
are identical to names of the points from the 'tempflow'-object of the given 'tempnetwork'-object.
and if the indices of each equivalence-class from the equivalence-relation of the given 'tempnetwork'-object,
are identical to names of the points from the 'tempflow'-object of the given 'tempnetwork'-object.
}
\description{
A function that checks,
that the names of the graphs in the graph-list of the given 'tempnetwork'-object
are identical to names of the points from the 'tempflow'-object of the given 'tempnetwork'-object.
and that the indices of each equivalence-class from the equivalence-relation of the given 'tempnetwork'-object,
are identical to names of the points from the 'tempflow'-object of the given 'tempnetwork'-object.
}
