% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tempnetwork-inputchecker.R
\name{check_tempflow_graphlist_index_equality.tempnetwork}
\alias{check_tempflow_graphlist_index_equality.tempnetwork}
\title{A function that checks,
if the names of the graphs in the graph-list of the given 'tempnetwork'-object
are identical to names of the points from the 'tempflow'-object of the given 'tempnetwork'-object.}
\usage{
\method{check_tempflow_graphlist_index_equality}{tempnetwork}(tempNetwork)
}
\arguments{
\item{tempNetwork}{A 'tempnetwork'-object.}
}
\value{
The boolean value TRUE,
if the names of the graphs in the graph-list of the given 'tempnetwork'-object
are identical to names of the points from the 'tempflow'-object of the given 'tempnetwork'-object.
}
\description{
A function that checks,
if the names of the graphs in the graph-list of the given 'tempnetwork'-object
are identical to names of the points from the 'tempflow'-object of the given 'tempnetwork'-object.
}
