% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tempnetwork-inputchecker.R
\name{check_tempflow_not_empty.tempnetwork}
\alias{check_tempflow_not_empty.tempnetwork}
\title{A function that checks, if the given 'tempflow'-object contains at least one point.}
\usage{
\method{check_tempflow_not_empty}{tempnetwork}(tempNetwork)
}
\arguments{
\item{tempFlow}{A 'tempflow'-object.}
}
\value{
The boolean value TRUE, if the given 'tempflow'-object contains at least one point.
}
\description{
A function that checks, if the given 'tempflow'-object contains at least one point.
}
