% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tempnetwork-inputchecker.R
\name{check_tempnetwork_completely.tempnetwork}
\alias{check_tempnetwork_completely.tempnetwork}
\title{A function that checks, all relevalnt 'tempnetwork'-conditions (see documentation of 'tempntwork' in general)}
\usage{
\method{check_tempnetwork_completely}{tempnetwork}(tempNetwork,
  throwError = TRUE)
}
\arguments{
\item{tempNetwork}{A 'tempnetwork'-object.}

\item{throwError}{If TRUE the respective errors will be thrown. Otherwise, a truth value will be returned.}
}
\value{
The boolean value TRUE,
if all relevalnt 'tempnetwork'-conditions
}
\description{
A function that checks, all relevalnt 'tempnetwork'-conditions (see documentation of 'tempntwork' in general)
}
