% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tempnetwork-inputchecker.R
\name{check_tvertices_exists.tempnetwork}
\alias{check_tvertices_exists.tempnetwork}
\title{A function that checks, if the equivalence-classes specified in "tVertexNameList" exist in the equivalence-relation of the given 'tempnetwork'-object.
If the equivalence-relation of the given 'tempnetwork'-object is NULL, it wull be check whether there exists a vertex with the same name in
the first graph of the graph-list from the given 'tempnetwork'-object.}
\usage{
\method{check_tvertices_exists}{tempnetwork}(tempNetwork, tVertexNameList)
}
\arguments{
\item{tempNetwork}{A 'tempnetwork'-object.}

\item{tVertexNameList}{A vector of strings representing the names of equivalence-classes in the equivalence-relation of the given 'tempnetwork'-object.}
}
\value{
If the equivalence-relation of the given 'tempnetwork'-object is NULL, it wull be check whether there exists a vertex with the same name in
the first graph of the graph-list from the given 'tempnetwork'-object.
Otherwise, the boolean value TRUE will be returned,
if the equivalence-classes specified in "tVertexNameList" exist in the equivalence-relation of the given 'tempnetwork'-object.
}
\description{
A function that checks, if the equivalence-classes specified in "tVertexNameList" exist in the equivalence-relation of the given 'tempnetwork'-object.
If the equivalence-relation of the given 'tempnetwork'-object is NULL, it wull be check whether there exists a vertex with the same name in
the first graph of the graph-list from the given 'tempnetwork'-object.
}
