% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputchecker.R
\name{check_vector_contains_na_or_or_nan_null}
\alias{check_vector_contains_na_or_or_nan_null}
\title{A function that checks, if the given vector contains either a NA, NaN or a NULL value.}
\usage{
check_vector_contains_na_or_or_nan_null(vector)
}
\arguments{
\item{vector1}{A arbitrary vector.}
}
\value{
The boolean value TRUE, if the given vector contains either a NA, NaN or a NULL value.
}
\description{
A function that checks, if the given vector contains either a NA, NaN or a NULL value.
}
