% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputchecker.R
\name{check_vector_equal_length}
\alias{check_vector_equal_length}
\title{A function that checks, if the two given vectors have equal length.}
\usage{
check_vector_equal_length(vector1, vector2)
}
\arguments{
\item{vector1}{A arbitrary vector.}

\item{vector2}{A arbitrary vector.}
}
\value{
The boolean value TRUE, if the two given vectors have equal length.
}
\description{
A function that checks, if the two given vectors have equal length.
}
