% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputchecker.R
\name{check_vector_equal_length_or_one}
\alias{check_vector_equal_length_or_one}
\title{A function that checks, if the 'checkVector' has either equal length as the 'refVector' or is of length 1.}
\usage{
check_vector_equal_length_or_one(checkVector, refVector)
}
\arguments{
\item{checkVector}{The vector to be checked.}

\item{refVector}{The reference vector.}
}
\value{
The boolean value TRUE, if the 'checkVector' has either equal length as the 'refVector' or is of length 1.
}
\description{
A function that checks, if the 'checkVector' has either equal length as the 'refVector' or is of length 1.
}
