% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tempflow-inputchecker.R
\name{check_weights_for_tempflow.tempflow}
\alias{check_weights_for_tempflow.tempflow}
\title{A function that checks, if the object assigned to the tempflow-field 'graph', has non-negative, numerical weights.}
\usage{
\method{check_weights_for_tempflow}{tempflow}(tempFlow,
  throwError = TRUE)
}
\arguments{
\item{tempFlow}{A 'tempflow'-object.}

\item{throwError}{If TRUE, violating the criteria above will result in the throwing of an error. Otherwise, FALSE will be returned.}
}
\value{
The boolean value TRUE, if the criteria above are satisfied. Otherwise, FALSE or an exception will be returned.
}
\description{
A function that checks, if the object assigned to the tempflow-field 'graph', has non-negative, numerical weights.
}
