% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tempflow-predicates.R
\name{comparable.tempflow}
\alias{comparable.tempflow}
\title{Is a function that takes a 'tempflow'-object and two point-names (or point-ids) and
tests whether there exists an either a path from "inpPoint1" to "inpPoint2" or a path from "inpPoint2" to "inpPoint1".
A point is not comparable to itself, without the existence of a loop.}
\usage{
\method{comparable}{tempflow}(tempFlow, inpPoint1, inpPoint2,
  safe = TRUE)
}
\arguments{
\item{tempFlow}{A 'tempflow'-object.}

\item{inpPoint1}{The point-name (or point-id).}

\item{inpPoint2}{The point-name (or point-id).}

\item{safe}{If TRUE, it will be tested, whether the given point-names (or point-ids) are present in the 'tempflow'-object.}
}
\value{
Is TRUE, if there exists an either a path from "inpPoint1" to "inpPoint2" or a path from "inpPoint2" to "inpPoint1".
}
\description{
Is a function that takes a 'tempflow'-object and two point-names (or point-ids) and
tests whether there exists an either a path from "inpPoint1" to "inpPoint2" or a path from "inpPoint2" to "inpPoint1".
A point is not comparable to itself, without the existence of a loop.
}
