% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tempnetwork-tempflow-predicates.R
\name{comparable.tempnetwork}
\alias{comparable.tempnetwork}
\title{Is a function that takes a 'tempnetwork'-object and two point-names (or point-ids) and
tests whether there exists an either a path from "inpPoint1" to "inpPoint2" or a path from "inpPoint2" to "inpPoint1" in 'tempflow'-object of the given 'tempnetwork'-object.
A point is not comparable to itself, without the existence of a loop.}
\usage{
\method{comparable}{tempnetwork}(tempNetwork, inpPoint1, inpPoint2,
  safe = TRUE)
}
\arguments{
\item{tempNetwork}{A 'tempnetwork'-object.}

\item{inpPoint1}{The point-name (or point-id).}

\item{inpPoint2}{The point-name (or point-id).}

\item{safe}{If TRUE, it will be tested, whether the given point-names (or point-ids) are present in the 'tempflow'-object of the given 'tempnetwork'-object.}
}
\value{
Is TRUE, if there exists an either a path from "inpPoint1" to "inpPoint2" or a path from "inpPoint2" to "inpPoint1" in 'tempflow'-object of the given 'tempnetwork'-object.
}
\description{
Is a function that takes a 'tempnetwork'-object and two point-names (or point-ids) and
tests whether there exists an either a path from "inpPoint1" to "inpPoint2" or a path from "inpPoint2" to "inpPoint1" in 'tempflow'-object of the given 'tempnetwork'-object.
A point is not comparable to itself, without the existence of a loop.
}
