% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tempnetwork-getter.R
\name{compute_equivrelation}
\alias{compute_equivrelation}
\title{Computes a default equivalence relation under the assumption that the vertex sets are identical.
That is, it starts by creating a list indexed by the names found in the parameter "defaultNameList".
Furthermore, each entry will be populated by a named list with it's names being the one's found in the "pointNameList" parameter
and with its values corresponding index of said list in the primary list.
E.g. compute_equivrelation(c("a","b","c","d"), c("t1","t2","t3")) ==
list("a"= list("t1"=a ,"t2"=a ,"t3"=a),"b"= list("t1"=b ,"t2"=b ,"t3"=b) ,"c"= list("t1"=c ,"t2"=c ,"t3"=c), "d"= list("t1"=d ,"t2"=d ,"t3"=d))}
\usage{
compute_equivrelation(defaultNameList, pointNameList)
}
\arguments{
\item{defaultNameList}{A vector containing strings representing  temporal vertex names, as well as the vertex names of each graph in some graph list.}

\item{pointNameList}{A vector containing strings representing the point names of an tempflow.}
}
\value{
An equivalence relation as described above.
}
\description{
Computes a default equivalence relation under the assumption that the vertex sets are identical.
That is, it starts by creating a list indexed by the names found in the parameter "defaultNameList".
Furthermore, each entry will be populated by a named list with it's names being the one's found in the "pointNameList" parameter
and with its values corresponding index of said list in the primary list.
E.g. compute_equivrelation(c("a","b","c","d"), c("t1","t2","t3")) ==
list("a"= list("t1"=a ,"t2"=a ,"t3"=a),"b"= list("t1"=b ,"t2"=b ,"t3"=b) ,"c"= list("t1"=c ,"t2"=c ,"t3"=c), "d"= list("t1"=d ,"t2"=d ,"t3"=d))
}
