% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tempflow-basics.R
\name{compute_tempdistance.tempflow}
\alias{compute_tempdistance.tempflow}
\title{Is a function that computes the temporal distance between two points in time with respect to a given 'tempflow'-object.
Moreover, it allows for the choice, whether weights should be considered in the computation or not.
The distance is based on the shortest path, that is if no shortest path can be found the distance will be infinite.}
\usage{
\method{compute_tempdistance}{tempflow}(tempFlow, startPoint, endPoint,
  considerWeights = TRUE, safe = TRUE)
}
\arguments{
\item{tempFlow}{A 'tempflow'-object.}

\item{startPoint}{A point-name (or point-id) representing the name of the start-point of the shortest path computation.}

\item{endPoint}{A point-name (or point-id) representing the name of the end-point of the shortest path computation.}

\item{considerWeights}{If TRUE weights will be considered in the shortest path computation, otherwise the distance will be given by
the number of steps between "startPoint" and endPoint"}

\item{safe}{If TRUE, it will be ensured that both start- and end-point exist in the 'tempflow'-object.}
}
\value{
The length or size of the shortest path. If no such path exists, it returns "Inf".
}
\description{
Is a function that computes the temporal distance between two points in time with respect to a given 'tempflow'-object.
Moreover, it allows for the choice, whether weights should be considered in the computation or not.
The distance is based on the shortest path, that is if no shortest path can be found the distance will be infinite.
}
