% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tempnetwork-basic.R
\name{construct_underlying_graph.tempnetwork}
\alias{construct_underlying_graph.tempnetwork}
\title{This function constructs a graph representation of the given 'tempnetwork'-object.
This is accomplished by renaming all vertices (as specified in the function "base_get_tempgraph_vertex_name"),
adding the attributes "vertexName", "pointName" and "tempVertexName"
according to their position in the equivalence-relation of the given 'tempnetwork'-object.
After this all edited graphs are merged (using igraph::disjoint_union).
Moreover, if connectGraphs is TRUE, for each equivalence-classe all contained vertices are connected in
the exact same manner as provided by the 'tempflow'-object of the given 'tempnetwork'-object.}
\usage{
\method{construct_underlying_graph}{tempnetwork}(tempNetwork,
  connectGraphs = TRUE, forceIGraph = FALSE)
}
\arguments{
\item{tempNetwork}{A 'tempnetwork'-object.}

\item{connectGraphs}{If TRUE, for each equivalence-classe all contained vertices are connected in
the exact same manner as provided by the 'tempflow'-object of the given 'tempnetwork'-object.}

\item{forceIGraph}{If TRUE, the function will always return an 'igraph'-object.}
}
\value{
If "forceIGraph" is TRUE or if the field "storeTempGraph" is FALSE, this function returns an 'igraph'-object as described above.
Otherwise, this function will return a 'tempnetwork'-object that is identical to the given 'tempnetwork'-object, but for the
fact that the field "tempGraph" contains an 'igraph'-object as described above.
}
\description{
This function constructs a graph representation of the given 'tempnetwork'-object.
This is accomplished by renaming all vertices (as specified in the function "base_get_tempgraph_vertex_name"),
adding the attributes "vertexName", "pointName" and "tempVertexName"
according to their position in the equivalence-relation of the given 'tempnetwork'-object.
After this all edited graphs are merged (using igraph::disjoint_union).
Moreover, if connectGraphs is TRUE, for each equivalence-classe all contained vertices are connected in
the exact same manner as provided by the 'tempflow'-object of the given 'tempnetwork'-object.
}
