% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tempflow-basics.R
\name{delete_points.tempflow}
\alias{delete_points.tempflow}
\title{Is a function that takes a 'tempflow'-object and a vector of point-names (or point-ids),
and removes the specified points from the 'tempflow'-object.}
\usage{
\method{delete_points}{tempflow}(tempFlow, deletePointList, safe = TRUE)
}
\arguments{
\item{tempFlow}{A 'tempflow'-object.}

\item{deletePointList}{A vector containing the point-names (or point-ids) of the points that ought to be removed.}

\item{safe}{If TRUE, it will be ensured that the specified point-names (or point-ids) are present in the given 'tempflow'-object.
Additionally, the usual tempflow condition are checked (e.g. duplicate names, cycles, ... ).}

\item{attrList}{A vector to specify other vertex attributes. If left NULL, no attributes (other than name) will be assigned to the newly introduced points in time.
Otherwise, this list has to be structured as required by the function igraph::add.vertices.}
}
\value{
A new 'tempflow'-object, constructed by removing the specified points from the given 'tempflow'-object.
}
\description{
Is a function that takes a 'tempflow'-object and a vector of point-names (or point-ids),
and removes the specified points from the 'tempflow'-object.
}
