% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tempnetwork-tempflow-basics.R
\name{delete_points.tempnetwork}
\alias{delete_points.tempnetwork}
\title{Is a function that takes a 'tempnetwork'-object and a vector of point-names (or point-ids),
and removes the specified points from the 'tempflow'-object of the given 'tempnetwork'-object.
Additionally, the corresponding entries in the "graphList" and in the "equivRel" of the given 'tempnetwork'-object
will be removed as well.}
\usage{
\method{delete_points}{tempnetwork}(tempNetwork, deletePointList,
  safe = TRUE)
}
\arguments{
\item{tempNetwork}{A 'tempnetwork'-object.}

\item{deletePointList}{A vector containing the point-names (or point-ids) of the points that ought to be removed.}

\item{safe}{If TRUE, it will be ensured that the specified point-names (or point-ids) are present in the given 'tempflow'-object.
Additionally, the usual tempflow and tempnetwork condition are checked.}

\item{attrList}{A vector to specify other vertex attributes. If left NULL, no attributes (other than name) will be assigned to the newly introduced points in time.
Otherwise, this list has to be structured as required by the function igraph::add.vertices.}
}
\value{
A new 'tempnetwork'-object, containing a new 'tempflow'-object constructed by removing the specified points from the 'tempflow'-object of the given 'tempnetwork'-object;
 containing a new "graphList" constructed by removing the specified entries from the existing "graphList" of the given 'tempnetwork'-object;
 containing a new "equivRel" constructed by removing the specified entries from the existing "equivRel" of the given 'tempnetwork'-object.
}
\description{
Is a function that takes a 'tempnetwork'-object and a vector of point-names (or point-ids),
and removes the specified points from the 'tempflow'-object of the given 'tempnetwork'-object.
Additionally, the corresponding entries in the "graphList" and in the "equivRel" of the given 'tempnetwork'-object
will be removed as well.
}
