% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tempflow-basics.R
\name{delete_steps.tempflow}
\alias{delete_steps.tempflow}
\title{Is a function that takes a 'tempflow'-object and two vectors of point-names (or point-ids)
and removes the steps from the 'tempflow'-object. The steps are characterised by the points in the "srcPointList"-vector and
the corresponding points in the "dstPointList"-vector,}
\usage{
\method{delete_steps}{tempflow}(tempFlow, srcPointList, dstPointList,
  safe = TRUE)
}
\arguments{
\item{tempFlow}{A 'tempflow'-object.}

\item{srcPointList}{A vector containing the point-names (or point-ids) of the start-points of the edges that ought to be removed.}

\item{dstPointList}{A vector containing the point-names (or point-ids) of the end-points of the edges that ought to be removed.}

\item{safe}{If TRUE, it will be ensured that the specified point-names (or point-ids) are present in the given 'tempflow'-object;
that the specified steps exist;
the "srcPointList"-vector, the "dstPointList"-vector are of the same length.
Additionally, the usual tempflow condition are checked (e.g. duplicate names, cycles, ... ).}
}
\value{
A new 'tempflow'-object, constructed by removing the specified edges from the given 'tempflow'-object.
}
\description{
Is a function that takes a 'tempflow'-object and two vectors of point-names (or point-ids)
and removes the steps from the 'tempflow'-object. The steps are characterised by the points in the "srcPointList"-vector and
the corresponding points in the "dstPointList"-vector,
}
