% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tempnetwork-tempflow-basics.R
\name{delete_steps.tempnetwork}
\alias{delete_steps.tempnetwork}
\title{Is a function that takes a 'tempnetwork'-object and two vectors of point-names (or point-ids)
and removes the steps from the 'tempflow'-object of the given 'tempnetwork'-object.
The steps are characterised by the points in the "srcPointList"-vector and
the corresponding points in the "dstPointList"-vector,}
\usage{
\method{delete_steps}{tempnetwork}(tempNetwork, srcPointList, dstPointList,
  safe = TRUE)
}
\arguments{
\item{tempNetwork}{A 'tempnetwork'-object.}

\item{srcPointList}{A vector containing the point-names (or point-ids) of the start-points of the edges that ought to be removed.}

\item{dstPointList}{A vector containing the point-names (or point-ids) of the end-points of the edges that ought to be removed.}

\item{safe}{If TRUE, it will be ensured that the specified point-names (or point-ids) are present in the 'tempflow'-object of the given 'tempnetwork'-object;
that the specified steps exist;
the "srcPointList"-vector, the "dstPointList"-vector are of the same length.
Additionally, the usual tempflow and tempnetwork condition are checked.}
}
\value{
A new 'tempnetwork'-object, containing a new 'tempflow'-object constructed by removing the specified edges from the 'tempflow'-object of the given 'tempnetwork'-object.
}
\description{
Is a function that takes a 'tempnetwork'-object and two vectors of point-names (or point-ids)
and removes the steps from the 'tempflow'-object of the given 'tempnetwork'-object.
The steps are characterised by the points in the "srcPointList"-vector and
the corresponding points in the "dstPointList"-vector,
}
