% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tempflow-basics.R
\name{delete_steps_from_id.tempflow}
\alias{delete_steps_from_id.tempflow}
\title{Given a 'tempflow'-object and a vector containing the current step-ids,
this function removes the specified steps from the given 'tempflow'-object.}
\usage{
\method{delete_steps_from_id}{tempflow}(tempFlow, stepIDList,
  safe = TRUE)
}
\arguments{
\item{tempFlow}{A 'tempflow'-object.}

\item{stepIDList}{A vector containing the step-ids of the steps which ought to be removed.}

\item{safe}{If TRUE, it will be ensured that the specified step-ids exists.
Additionally, the usual tempflow condition are checked (e.g. duplicate names, cycles, ... ).}
}
\value{
A new 'tempflow'-object, constructed by removing the specified edges from the given 'tempflow'-object.
}
\description{
Given a 'tempflow'-object and a vector containing the current step-ids,
this function removes the specified steps from the given 'tempflow'-object.
}
