% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tempnetwork-tempflow-basics.R
\name{delete_steps_from_id.tempnetwork}
\alias{delete_steps_from_id.tempnetwork}
\title{Given a 'tempnetwork'-object and a vector containing the current step-ids,
this function removes the specified steps from the 'tempflow'-object of the given 'tempnetwork'-object.}
\usage{
\method{delete_steps_from_id}{tempnetwork}(tempNetwork, stepIDList,
  safe = TRUE)
}
\arguments{
\item{tempNetwork}{A 'tempnetwork'-object.}

\item{stepIDList}{A vector containing the step-ids of the steps which ought to be removed.}

\item{safe}{If TRUE, it will be ensured that the specified step-ids exists.
Additionally, the usual tempflow and tempnetwork condition are checked.}
}
\value{
A new 'tempnetwork'-object, containing a new 'tempflow'-object constructed by removing the specified edges from the 'tempflow'-object of the given 'tempnetwork'-object.
}
\description{
Given a 'tempnetwork'-object and a vector containing the current step-ids,
this function removes the specified steps from the 'tempflow'-object of the given 'tempnetwork'-object.
}
