% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tempnetwork-getter.R
\name{eR.tempnetwork}
\alias{eR.tempnetwork}
\title{Returns the equivalence-relation of the given 'tempnetwork'-object.}
\usage{
\method{eR}{tempnetwork}(tempNetwork, computeEquivRel = FALSE)
}
\arguments{
\item{tempNetwork}{A 'tempnetwork'-object.}
}
\value{
The equivalence-relation of the given 'tempnetwork'-object.
}
\description{
Returns the equivalence-relation of the given 'tempnetwork'-object.
}
