% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tempflow-getter.R
\name{G.tempflow}
\alias{G.tempflow}
\title{Returns the graph underlying the given 'tempflow'-object.}
\usage{
\method{G}{tempflow}(tempFlow)
}
\arguments{
\item{tempFlow}{A 'tempflow'-object.}
}
\value{
The 'igraph'-object underlying, the given 'tempflow'-object.
}
\description{
Returns the graph underlying the given 'tempflow'-object.
}
