% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tempnetwork-tempflow-getter.R
\name{G.tempnetwork}
\alias{G.tempnetwork}
\title{Returns the graph underlying the 'tempflow'-object of the given 'tempnetwork'-object.}
\usage{
\method{G}{tempnetwork}(tempNetwork)
}
\arguments{
\item{tempNetwork}{A 'tempnetwork'-object.}
}
\value{
The 'igraph'-object underlying, the 'tempflow'-object of the given 'tempnetwork'-object.
}
\description{
Returns the graph underlying the 'tempflow'-object of the given 'tempnetwork'-object.
}
