% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tempflow-getter.R
\name{get_point_ids.tempflow}
\alias{get_point_ids.tempflow}
\title{Given a vector of point-names (or a vector of point-ids) this function returns a vector containing the corresponding point-ids
as defined in (the 'igraph'-object underlying) the given 'tempflow'-object.
If one of the given point-names (or point-ids) is not present and if 'safe' has the value TRUE, an error will be thrown.}
\usage{
\method{get_point_ids}{tempflow}(tempFlow, inpPointList = tP(tempFlow),
  safe = TRUE)
}
\arguments{
\item{tempFlow}{A 'tempflow'-object.}

\item{inpPointList}{Either a vector of strings representing point-names or
a vector of integers representing a point-ids.}

\item{safe}{If TRUE, it will be tested, whether the given point-names (or point-ids) are present in the 'tempflow'-object.}
}
\value{
The point-ids of the points specified in "inpPointList", as currently defined in the 'tempflow'-object "tempFlow".
}
\description{
Given a vector of point-names (or a vector of point-ids) this function returns a vector containing the corresponding point-ids
as defined in (the 'igraph'-object underlying) the given 'tempflow'-object.
If one of the given point-names (or point-ids) is not present and if 'safe' has the value TRUE, an error will be thrown.
}
