% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tempnetwork-tempflow-getter.R
\name{get_point_name.tempnetwork}
\alias{get_point_name.tempnetwork}
\title{Given a point-id (or a point-name) this function returns the corresponding point-name
as defined in (the 'igraph'-object underlying) the 'tempflow'-object of the given 'tempnetwork'-object.
If the given point-id or point-name is not present and if 'safe' has the value TRUE, an error will be thrown.}
\usage{
\method{get_point_name}{tempnetwork}(tempNetwork, inpPoint, safe = TRUE)
}
\arguments{
\item{tempNetwork}{A 'tempnetwork'-object.}

\item{inpPoint}{Either a string representing the point-name or
an integers representing the point-id.}

\item{safe}{If TRUE, it will be tested, whether the given point-id (or point-name) is present in the 'tempflow'-object of the given 'tempnetwork'-object.}
}
\value{
The point-name of the point specified in "inpPoint" as currently defined in the 'tempflow'-object of the given 'tempnetwork'-object.
}
\description{
Given a point-id (or a point-name) this function returns the corresponding point-name
as defined in (the 'igraph'-object underlying) the 'tempflow'-object of the given 'tempnetwork'-object.
If the given point-id or point-name is not present and if 'safe' has the value TRUE, an error will be thrown.
}
