% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tempnetwork-tempflow-getter.R
\name{get_step_attr_list_from_id.tempnetwork}
\alias{get_step_attr_list_from_id.tempnetwork}
\title{This function returns the attribute list of the specified steps in the 'tempflow'-object of the given 'tempnetwork'-object.}
\usage{
\method{get_step_attr_list_from_id}{tempnetwork}(tempNetwork,
  stepIDList = tS(tempNetwork), safe = TRUE)
}
\arguments{
\item{tempNetwork}{A 'tempnetwork'-object.}

\item{stepIDList}{A vector containing the edge-ids of the desired edges, as currently defined in (the 'igraph'-object underlying) the 'tempflow'-object of the given 'tempnetwork'-object.}

\item{safe}{If TRUE, it will be tested, whether the specified edges are present in the 'tempnetwork'-object of the given 'tempnetwork'-object.}
}
\value{
This function returns the attribute list of the specified steps from the 'tempflow'-object of the given 'tempnetwork'-object.
}
\description{
This function returns the attribute list of the specified steps in the 'tempflow'-object of the given 'tempnetwork'-object.
}
