% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tempnetwork-tempflow-getter.R
\name{get_step_ids.tempnetwork}
\alias{get_step_ids.tempnetwork}
\title{This function returns the edge-ids (as currently defined in the 'igraph'-object underlying the 'tempflow'-object of the given 'tempnetwork'-object),
of the edges which are characterised by their start- and endpoints that are specified by the input vectors
srcPointList and dstPointList respectively, both of which are containing point-names (or point-ids).}
\usage{
\method{get_step_ids}{tempnetwork}(tempNetwork, srcPointList = NULL,
  dstPointList = NULL, safe = TRUE)
}
\arguments{
\item{tempNetwork}{A 'tempnetwork'-object.}

\item{srcPointList}{The vector containing the point-names (or point-ids) representing the start-points of the desired edges.}

\item{dstPointList}{The vector containing the point-names (or point-ids) representing the end-points of the desired edges.}

\item{safe}{If TRUE, it will be tested, whether the specified point-names (or point-ids) are present in the 'tempflow'-object of the given 'tempnetwork'-object;
whether the vectors have the same length; whether the specified edges are present in the 'tempflow'-object of the given 'tempnetwork'-object.}
}
\value{
A vector of the current step-ids, as defined in the 'igraph'-object underlying the 'tempflow'-object of the given 'tempnetwork'-object.
}
\description{
This function returns the edge-ids (as currently defined in the 'igraph'-object underlying the 'tempflow'-object of the given 'tempnetwork'-object),
of the edges which are characterised by their start- and endpoints that are specified by the input vectors
srcPointList and dstPointList respectively, both of which are containing point-names (or point-ids).
}
