% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tempnetwork-tempflow-getter.R
\name{get_step_point_list.tempnetwork}
\alias{get_step_point_list.tempnetwork}
\title{This function returns a named list. The key 'startPointList' references the start-points of the specified steps,
while the key 'endPointList' references their end-points.}
\usage{
\method{get_step_point_list}{tempnetwork}(tempNetwork,
  srcPointList = NULL, dstPointList = NULL, safe = TRUE)
}
\arguments{
\item{tempNetwork}{A 'tempnetwork'-object.}

\item{srcPointList}{The vector containing the point-names (or point-ids) representing the start-points of the desired edges.}

\item{dstPointList}{The vector containing the point-names (or point-ids) representing the end-points of the desired edges.}

\item{safe}{If TRUE, it will be tested, whether the specified point-names (or point-ids) are present in the 'tempflow'-object of the given 'tempnetwork'-object;
whether the vectors have the same length; whether the specified edges are present in the 'tempflow'-object of the given 'tempnetwork'-object.}
}
\value{
A list with the keys 'startPointList' and 'endPointList' containing the respective start- and end-points of the given steps.
}
\description{
This function returns a named list. The key 'startPointList' references the start-points of the specified steps,
while the key 'endPointList' references their end-points.
}
