% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tempflow-getter.R
\name{get_step_point_list_from_id.tempflow}
\alias{get_step_point_list_from_id.tempflow}
\title{This function returns a named list. The key 'startPointList' references the start-points of the specified steps,
while the key 'endPointList' references their end-points.}
\usage{
\method{get_step_point_list_from_id}{tempflow}(tempFlow,
  stepIDList = tS(tempFlow), safe = TRUE)
}
\arguments{
\item{tempFlow}{A 'tempflow'-object.}

\item{stepIDList}{A vector containing the edge-ids of the desired edges, as currently defined in (the 'igraph'-object underlying) the given 'tempflow'-object.}

\item{safe}{If TRUE, it will be tested, whether the specified edges are present in the given 'tempflow'-object.}
}
\value{
A list with the keys 'startPointList' and 'endPointList' containing the respective start- and end-points of the given steps.
}
\description{
This function returns a named list. The key 'startPointList' references the start-points of the specified steps,
while the key 'endPointList' references their end-points.
}
