% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tempnetwork-tempflow-getter.R
\name{get_step_point_list_from_id.tempnetwork}
\alias{get_step_point_list_from_id.tempnetwork}
\title{This function returns a named list. The key 'startPointList' references the start-points of the specified steps,
while the key 'endPointList' references their end-points.}
\usage{
\method{get_step_point_list_from_id}{tempnetwork}(tempNetwork,
  stepIDList = tS(tempNetwork), safe = TRUE)
}
\arguments{
\item{tempNetwork}{A 'tempnetwork'-object.}

\item{stepIDList}{A vector containing the edge-ids of the desired edges, as currently defined in (the 'igraph'-object underlying) the 'tempflow'-object of the given 'tempnetwork'-object.}

\item{safe}{If TRUE, it will be tested, whether the specified edges are present in the 'tempflow'-object of the given 'tempnetwork'-object.}
}
\value{
A list with the keys 'startPointList' and 'endPointList' containing the respective start- and end-points of the given steps.
}
\description{
This function returns a named list. The key 'startPointList' references the start-points of the specified steps,
while the key 'endPointList' references their end-points.
}
