% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tempflow-getter.R
\name{get_step_weights.tempflow}
\alias{get_step_weights.tempflow}
\title{This function returns the edge-weights, as currently defined in (the 'igraph'-object underlying) the given 'tempflow'-object,
of the edges which are characterised by their start- and end-points that are specified by the input vectors,
"srcPointList" and "dstPointList" respectively, both of which are containing point-names (or point-ids).}
\usage{
\method{get_step_weights}{tempflow}(tempFlow, srcPointList = NULL,
  dstPointList = NULL, safe = TRUE)
}
\arguments{
\item{tempFlow}{A 'tempflow'-object.}

\item{srcPointList}{The vector containing the point-names (or point-ids) representing the start-points of the desired edges.}

\item{dstPointList}{The vector containing the point-names (or point-ids) representing the end-points of the desired edges.}

\item{safe}{If TRUE, it will be tested, whether the specified point-names (or point-ids) are present in the given 'tempflow'-object;
whether the vectors have the same length; whether the specified edges are present in the given 'tempflow'-object.}
}
\value{
A vector containing the step-weights of the specified steps, as defined in (the 'igraph'-object underlying) the given 'tempflow'-object.
}
\description{
This function returns the edge-weights, as currently defined in (the 'igraph'-object underlying) the given 'tempflow'-object,
of the edges which are characterised by their start- and end-points that are specified by the input vectors,
"srcPointList" and "dstPointList" respectively, both of which are containing point-names (or point-ids).
}
