% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tempflow-getter.R
\name{get_step_weights_from_id.tempflow}
\alias{get_step_weights_from_id.tempflow}
\title{This function returns the edge-weights, as currently defined in (the 'igraph'-object underlying) the given 'tempflow'-object,
of the edges which are specified by a vector containing their edge-ids
(as currently defined in (the 'igraph'-object underlying) the given 'tempflow'-object).}
\usage{
\method{get_step_weights_from_id}{tempflow}(tempFlow,
  stepIDList = tS(tempFlow), safe = TRUE)
}
\arguments{
\item{tempFlow}{A 'tempflow'-object.}

\item{stepIDList}{A vector containing the edge-ids of the desired edges, as currently defined in (the 'igraph'-object underlying) the given 'tempflow'-object.}

\item{safe}{If TRUE, it will be tested, whether the specified edges are present in the given 'tempflow'-object.}
}
\value{
A vector of the current step-weights, as defined in (the 'igraph'-object underlying) the given 'tempflow'-object.
}
\description{
This function returns the edge-weights, as currently defined in (the 'igraph'-object underlying) the given 'tempflow'-object,
of the edges which are specified by a vector containing their edge-ids
(as currently defined in (the 'igraph'-object underlying) the given 'tempflow'-object).
}
