% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tempnetwork-tempflow-getter.R
\name{get_step_weights_from_id.tempnetwork}
\alias{get_step_weights_from_id.tempnetwork}
\title{This function returns the edge-weights, as currently defined in (the 'igraph'-object underlying) the 'tempflow'-object of the given 'tempnetwork'-object,
of the edges which are specified by a vector containing their edge-ids
(as currently defined in (the 'igraph'-object underlying) the 'tempflow'-object of the given 'tempnetwork'-object).}
\usage{
\method{get_step_weights_from_id}{tempnetwork}(tempNetwork,
  stepIDList = tS(tempNetwork), safe = TRUE)
}
\arguments{
\item{tempNetwork}{A 'tempnetwork'-object.}

\item{stepIDList}{A vector containing the edge-ids of the desired edges, as currently defined in (the 'igraph'-object underlying) the 'tempflow'-object of the given 'tempnetwork'-object.}

\item{safe}{If TRUE, it will be tested, whether the specified edges are present in the 'tempflow'-object of the given 'tempnetwork'-object.}
}
\value{
A vector of the current step-weights, as defined in (the 'igraph'-object underlying) the 'tempflow'-object of the given 'tempnetwork'-object.
}
\description{
This function returns the edge-weights, as currently defined in (the 'igraph'-object underlying) the 'tempflow'-object of the given 'tempnetwork'-object,
of the edges which are specified by a vector containing their edge-ids
(as currently defined in (the 'igraph'-object underlying) the 'tempflow'-object of the given 'tempnetwork'-object).
}
