% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tempnetwork-getter.R
\name{get_tempgraph_vertex_names.tempnetwork}
\alias{get_tempgraph_vertex_names.tempnetwork}
\title{Using the strings in the tVertexNameList-vector and the strings in the pointNameList-vector function computes the names of the corresponding vertices
in the tempGraph representation of the given 'tempnetwork'-object.}
\usage{
\method{get_tempgraph_vertex_names}{tempnetwork}(tempNetwork,
  tVertexNameList, pointNameList, safe = TRUE)
}
\arguments{
\item{tempNetwork}{A 'tempnetwork'-object.}

\item{tVertexNameList}{A vector of strings corresponding with the names of the equivalence classes specified in the "equigRel"  of the given 'tempnetwork'-object.}

\item{pointNameList}{A vector of strings corresponding with the names of points in the 'tempflow'-object of the given 'tempnetwork'-object.}

\item{safe}{If TRUE, it will be ensured that every equivalence class specified in the "tVertexNameList"-vector and every point in the "pointNameList"-vector exist
in their respective structures found in the given 'tempnetwork'-object.}
}
\value{
A vector of strings representing the names of the corresponding vertices in the tempGraph representation of the given 'tempnetwork'-object.
}
\description{
Using the strings in the tVertexNameList-vector and the strings in the pointNameList-vector function computes the names of the corresponding vertices
in the tempGraph representation of the given 'tempnetwork'-object.
}
