% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tempnetwork-getter.R
\name{get_vertexname.tempnetwork}
\alias{get_vertexname.tempnetwork}
\title{Queries the equivalence relation to obtain the name of instance of the tVertex "tVertexName" at the point "pointName"}
\usage{
\method{get_vertexname}{tempnetwork}(tempNetwork, tVertexName, pointName,
  safe = TRUE)
}
\arguments{
\item{tempNetwork}{A 'tempnetwork'-object.}

\item{tVertexName}{A string corresponding with the name of a tVertex specified in the "equigRel"  of the given 'tempnetwork'-object.}

\item{pointName}{A string corresponding with the name of a point in the 'tempflow'-object of the given 'tempnetwork'-object.}

\item{safe}{If TRUE, it will be ensured that "tVertexName" and "pointName" exist.}
}
\value{
A string corresponding to the name of a vertex in the graph indexed by "pointName", that was assigned membership to the
equivalence class "tVertexName" by the "equivRel" of the given "tempnetwork"-object.
}
\description{
Queries the equivalence relation to obtain the name of instance of the tVertex "tVertexName" at the point "pointName"
}
