% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tempnetwork-getter.R
\name{get_vertexnames.tempnetwork}
\alias{get_vertexnames.tempnetwork}
\title{Queries the equivalence relation to obtain the name of instance of the tVertex "tVertexName" at the point "pointName"}
\usage{
\method{get_vertexnames}{tempnetwork}(tempNetwork, tVertexNameList,
  pointNameList, safe = TRUE)
}
\arguments{
\item{tempNetwork}{A 'tempnetwork'-object.}

\item{safe}{If TRUE, it will be ensured that all entries in "tVertexNameList" and "pointNameList" exist in their respective data structures.}

\item{tVertexName}{A vector of strings corresponding with the names of equivalence-classes in the equivalence-relation of the given 'tempnetwork'-object.}

\item{pointName}{A vector of strings corresponding with the names of points in the 'tempflow'-object of the given 'tempnetwork'-object.}
}
\value{
A vector of String corresponding to the names of vertices in the graphs indexed by the entries in the "pointNameList"-vector, that was assigned membership to the
equivalence class "tVertexNameList" by the "equivRel" of the given "tempnetwork"-object.
}
\description{
Queries the equivalence relation to obtain the name of instance of the tVertex "tVertexName" at the point "pointName"
}
