% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tempnetwork-getter.R
\name{gL.tempnetwork}
\alias{gL.tempnetwork}
\title{Returns the graph-list of the given 'tempnetwork'-object.}
\usage{
\method{gL}{tempnetwork}(tempNetwork)
}
\arguments{
\item{tempNetwork}{A 'tempnetwork'-object.}
}
\value{
The graph-list of the given 'tempnetwork'-object.
}
\description{
Returns the graph-list of the given 'tempnetwork'-object.
}
