% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tempnetwork-inputchecker.R
\name{graphlist_cleaner}
\alias{graphlist_cleaner}
\title{Any graph containing a vertex with a name that is not a string, will be
coerced into a string.
Moreover, if "setDefaultGraphNames" is TRUE, any graph in the given graph-list containing a vertex without a name, will be replaced by a graph where the name of each
vertex is a string corresponding to its index in the respective 'igraph'-object.}
\usage{
graphlist_cleaner(graphList, setDefaultGraphNames = TRUE)
}
\arguments{
\item{graphList}{A named list of 'igraph'-objects.}

\item{setDefaultGraphNames}{If TRUE, any graph in the given graph-list containing a vertex without a name, will be replaced by a graph where the name of each
vertex is a string corresponding to its index in the respective 'igraph'-object.
coerced into a string.}
}
\value{
A graph-list as described above.
}
\description{
Any graph containing a vertex with a name that is not a string, will be
coerced into a string.
Moreover, if "setDefaultGraphNames" is TRUE, any graph in the given graph-list containing a vertex without a name, will be replaced by a graph where the name of each
vertex is a string corresponding to its index in the respective 'igraph'-object.
}
