% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tempflow-predicates.R
\name{greater_eq.tempflow}
\alias{greater_eq.tempflow}
\title{Is a function that takes a 'tempflow'-object and two point-names (or point-ids) and
tests whether "inpPoint1" is greater than or equal to "inpPoint2" (given the order induced by the
structure of the 'tempflow'-object).}
\usage{
\method{greater_eq}{tempflow}(tempFlow, inpPoint1, inpPoint2,
  safe = TRUE)
}
\arguments{
\item{tempFlow}{A 'tempflow'-object.}

\item{inpPoint1}{A point-name (or point-id).}

\item{inpPoint2}{A point-name (or point-id).}

\item{safe}{If TRUE, it will be tested, whether the given point-names (or point-ids) are present in the 'tempflow'-object.}
}
\value{
Is TRUE, if "inpPoint1" is strictly greater than "inpPoint2" (given the order induced by the
structure of the 'tempflow'-object).
}
\description{
Is a function that takes a 'tempflow'-object and two point-names (or point-ids) and
tests whether "inpPoint1" is greater than or equal to "inpPoint2" (given the order induced by the
structure of the 'tempflow'-object).
}
