% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper_functions.R
\name{igE}
\alias{igE}
\title{Returns the edges of the given 'igraph'-object.}
\usage{
igE(graph)
}
\arguments{
\item{graph}{A 'igraph'-object.}
}
\value{
The vertices of the given 'igraph'-object.
}
\description{
Returns the edges of the given 'igraph'-object.
}
