% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper_functions.R
\name{igV}
\alias{igV}
\title{Returns the vertices of the given 'igraph'-object.}
\usage{
igV(graph)
}
\arguments{
\item{graph}{A 'igraph'-object.}
}
\value{
The vertices of the given 'igraph'-object.
}
\description{
Returns the vertices of the given 'igraph'-object.
}
