% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tempflow-basics.R
\name{induce_flow.tempflow}
\alias{induce_flow.tempflow}
\title{Is a function that takes a 'tempflow'-object and a vector of point-names (or point-ids)
and restricts the given flow to the sub-flow induced by the vector of points.
That is, the underlying 'igraph'-object will be restricted to a graph containing only the vertices
corresponding to the specified points.}
\usage{
\method{induce_flow}{tempflow}(tempFlow, inpPointList, safe = TRUE)
}
\arguments{
\item{tempFlow}{A 'tempflow'-object.}

\item{inpPointList}{A vector containing point-names (or point-ids).}

\item{safe}{If TRUE, it will be ensured that all the specified points exist in the 'tempflow'-object and
that all relevant 'tempflow'-conditions are uphold.}
}
\value{
A new 'tempflow'-object, obtained by restricting the underlying 'igraph'-object
to the vertices corresponding to the points specified in the "inpPointList"-vector.
}
\description{
Is a function that takes a 'tempflow'-object and a vector of point-names (or point-ids)
and restricts the given flow to the sub-flow induced by the vector of points.
That is, the underlying 'igraph'-object will be restricted to a graph containing only the vertices
corresponding to the specified points.
}
