% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tempnetwork-tempflow-basics.R
\name{induce_flow.tempnetwork}
\alias{induce_flow.tempnetwork}
\title{Is a function that takes a 'tempnetwork'-object and a vector of point-names (or point-ids)
and restricts the underlying flow to the sub-flow induced by the vector of points.
That is, the 'igraph'-object underlying the 'tempflow'-object will be restricted to a graph containing only the vertices
corresponding to the specified points.}
\usage{
\method{induce_flow}{tempnetwork}(tempNetwork, inpPointList, safe = TRUE)
}
\arguments{
\item{tempNetwork}{A 'tempnetwork'-object.}

\item{inpPointList}{A vector containing point-names (or point-ids).}

\item{safe}{If TRUE, it will be ensured that all the specified points exist in the 'tempflow'-object of the given 'tempnetwork'-object and
that all relevant 'tempflow'-conditions and 'tempnetwork'-conditions are uphold.}
}
\value{
A new 'tempnetwork'-object, containing a new 'tempflow'-object obtained by restricting the underlying 'igraph'-object
to the vertices corresponding to the points specified in the "inpPointList"-vector;
 containing a new "graphList" constructed by restricting the existing "graphList" to those entries corresponding to the points in the  new 'tempflow'-object;
 containing a new "equivRel" constructed by restricting the existing "equivRel" to those entries corresponding to the points in the  new 'tempflow'-object.
}
\description{
Is a function that takes a 'tempnetwork'-object and a vector of point-names (or point-ids)
and restricts the underlying flow to the sub-flow induced by the vector of points.
That is, the 'igraph'-object underlying the 'tempflow'-object will be restricted to a graph containing only the vertices
corresponding to the specified points.
}
