% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tempflow-predicates.R
\name{is_branch_globally_homogeneous.tempflow}
\alias{is_branch_globally_homogeneous.tempflow}
\title{Is a function that checks whether the 'tempflow'-object, between two points is globally homogeneous.
That is, whether every path between the two points in time has the same size (or length).}
\usage{
\method{is_branch_globally_homogeneous}{tempflow}(tempFlow, startPoint,
  endPoint, considerWeights = TRUE, considerLoops = TRUE,
  safe = TRUE)
}
\arguments{
\item{tempFlow}{A 'tempflow'-object.}

\item{considerWeights}{If TRUE weights will be considered in the computation of the shortest paths,
otherwise the distance will be equal to the number of steps.}

\item{safe}{If TRUE, it will be tested, whether the given point-names (or point-ids) are present in the 'tempflow'-object.}

\item{srcPoint}{The point-name (or point-id) of the start-point of the computation.}

\item{trgPoint}{The point-name (or point-id) of the end-point of the computation.}

\item{checkForLoops}{If FALSE, loops will be ignored in the computation. That is, only paths without loops will be considered.}
}
\value{
A truth value indicating whether the flow between given two points is globally homogeneous.
}
\description{
Is a function that checks whether the 'tempflow'-object, between two points is globally homogeneous.
That is, whether every path between the two points in time has the same size (or length).
}
