% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tempnetwork-tempflow-predicates.R
\name{is_branch_globally_homogeneous.tempnetwork}
\alias{is_branch_globally_homogeneous.tempnetwork}
\title{Is a function that checks whether the 'tempflow'-object of the given 'tempnetwork'-object is globally homogeneous in between the two given points.
That is, whether every path between the two points in time has the same size (or length).}
\usage{
\method{is_branch_globally_homogeneous}{tempnetwork}(tempNetwork,
  startPoint, endPoint, considerWeights = TRUE, considerLoops = TRUE,
  safe = TRUE)
}
\arguments{
\item{tempNetwork}{A 'tempnetwork'-object.}

\item{considerWeights}{If TRUE weights will be considered in the computation of the shortest paths,
otherwise the distance will be equal to the number of steps.}

\item{safe}{If TRUE, it will be tested, whether the given point-names (or point-ids) are present in the 'tempflow'-object of the given 'tempnetwork'-object.}

\item{srcPoint}{The point-name (or point-id) of the start-point of the computation.}

\item{trgPoint}{The point-name (or point-id) of the end-point of the computation.}

\item{checkForLoops}{If FALSE, loops will be ignored in the computation. That is, only paths without loops will be considered.}
}
\value{
A truth value indicating whether the undelying flow is globally homogeneous inbetween the two given points.
}
\description{
Is a function that checks whether the 'tempflow'-object of the given 'tempnetwork'-object is globally homogeneous in between the two given points.
That is, whether every path between the two points in time has the same size (or length).
}
