% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tempflow-predicates.R
\name{is_branch_locally_homogeneous.tempflow}
\alias{is_branch_locally_homogeneous.tempflow}
\title{Is a function that checks whether the 'tempflow'-object between two points is locally homogeneous.
That is, whether every step (excluding loops) has the same weight.}
\usage{
\method{is_branch_locally_homogeneous}{tempflow}(tempFlow, startPoint,
  endPoint, safe = TRUE)
}
\arguments{
\item{tempFlow}{A 'tempflow'-object.}

\item{startPoint}{A point-name (or point-id).}

\item{endPoint}{A point-name (or point-id).}

\item{safe}{If TRUE, it will be tested, whether the given point-names (or point-ids) are present in the 'tempflow'-object.}
}
\value{
A truth value indicating whether the flow between the given two points is locally homogeneous.
}
\description{
Is a function that checks whether the 'tempflow'-object between two points is locally homogeneous.
That is, whether every step (excluding loops) has the same weight.
}
