% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tempnetwork-tempflow-predicates.R
\name{is_branch_locally_homogeneous.tempnetwork}
\alias{is_branch_locally_homogeneous.tempnetwork}
\title{Is a function that checks whether the 'tempflow'-object of the given 'tempnetwork'-object is locally homogeneous inbetween two given points .
That is, whether every step (excluding loops) has the same weight.}
\usage{
\method{is_branch_locally_homogeneous}{tempnetwork}(tempNetwork,
  startPoint, endPoint, safe = TRUE)
}
\arguments{
\item{tempNetwork}{A 'tempnetwork'-object.}

\item{startPoint}{A point-name (or point-id).}

\item{endPoint}{A point-name (or point-id).}

\item{safe}{If TRUE, it will be tested, whether the given point-names (or point-ids) are present in the 'tempflow'-object of the given 'tempnetwork'-object.}
}
\value{
A truth value indicating whether the underlying flow between the given two points is locally homogeneous.
}
\description{
Is a function that checks whether the 'tempflow'-object of the given 'tempnetwork'-object is locally homogeneous inbetween two given points .
That is, whether every step (excluding loops) has the same weight.
}
