% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tempflow-predicates.R
\name{is_globally_homogeneous.tempflow}
\alias{is_globally_homogeneous.tempflow}
\title{Is a function that checks whether the 'tempflow'-object is globally homogeneous.
That is, whether every path between any two points in time has the same size (or length).}
\usage{
\method{is_globally_homogeneous}{tempflow}(tempFlow,
  considerWeights = TRUE, considerLoops = TRUE)
}
\arguments{
\item{tempFlow}{A 'tempflow'-object.}

\item{considerWeights}{If TRUE weights will be considered in the computation of the shortest paths,
otherwise the distance will be equal to the number of steps.}

\item{checkForLoops}{If FALSE, loops will be ignored in the computation. That is, only paths without loops will be considered.}
}
\value{
A truth value indicating whether the flow is globally homogeneous.
}
\description{
Is a function that checks whether the 'tempflow'-object is globally homogeneous.
That is, whether every path between any two points in time has the same size (or length).
}
