% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tempnetwork-tempflow-predicates.R
\name{is_globally_homogeneous.tempnetwork}
\alias{is_globally_homogeneous.tempnetwork}
\title{Is a function that checks whether the 'tempflow'-object of the given 'tempnetwork'-object is globally homogeneous.
That is, whether every path between any two points in time has the same size (or length).}
\usage{
\method{is_globally_homogeneous}{tempnetwork}(tempNetwork,
  considerWeights = TRUE, considerLoops = TRUE)
}
\arguments{
\item{tempNetwork}{A 'tempnetwork'-object.}

\item{considerWeights}{If TRUE weights will be considered in the computation of the shortest paths,
otherwise the distance will be equal to the number of steps.}

\item{checkForLoops}{If FALSE, loops will be ignored in the computation. That is, only paths without loops will be considered.}
}
\value{
A truth value indicating whether the underlying flow is globally homogeneous.
}
\description{
Is a function that checks whether the 'tempflow'-object of the given 'tempnetwork'-object is globally homogeneous.
That is, whether every path between any two points in time has the same size (or length).
}
