% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tempflow-predicates.R
\name{is_locally_homogeneous.tempflow}
\alias{is_locally_homogeneous.tempflow}
\title{Is a function that checks whether the 'tempflow'-object is locally homogeneous.
That is, whether every step (excluding loops) has the same weight.}
\usage{
\method{is_locally_homogeneous}{tempflow}(tempFlow)
}
\arguments{
\item{tempFlow}{A 'tempflow'-object.}
}
\value{
A truth value indicating whether the flow is locally homogeneous.
}
\description{
Is a function that checks whether the 'tempflow'-object is locally homogeneous.
That is, whether every step (excluding loops) has the same weight.
}
