% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tempflow-getter.R
\name{plot.tempflow}
\alias{plot.tempflow}
\title{Is a wrapper function of the normal plot function. Allowing to plot a tempflow objec like a graph.}
\usage{
\method{plot}{tempflow}(tempFlow, weighted = TRUE, ...)
}
\arguments{
\item{tempFlow}{A 'tempflow'-object.}

\item{weighted}{If TRUE the weights of the steps in time will be plotted.}
}
\value{
Returns a plot of the 'tempflow'-object
}
\description{
Is a wrapper function of the normal plot function. Allowing to plot a tempflow objec like a graph.
}
