% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tempnetwork-tempflow-getter.R
\name{plot.tempnetwork}
\alias{plot.tempnetwork}
\title{Is a wrapper function of the normal plot function. Allowing to plot a tempflow objec like a graph.}
\usage{
\method{plot}{tempnetwork}(tempNetwork, weighted = TRUE, levels = TRUE,
  ...)
}
\arguments{
\item{tempNetwork}{A 'tempnetwork'-object.}

\item{weighted}{If TRUE the weights of the steps in time will be plotted.}
}
\value{
Returns a plot of the 'tempnetwork'-object
}
\description{
Is a wrapper function of the normal plot function. Allowing to plot a tempflow objec like a graph.
}
