% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tempnetwork-tempflow-predicates.R
\name{reachable.tempnetwork}
\alias{reachable.tempnetwork}
\title{Is a function that takes a 'tempnetwork'-object and two point-names (or point-ids) and
tests whether one point ("endPoint") can be reached from the other ("startPoint") in the 'tempflow'-object of the given 'tempnetwork'-object.
A point can not reach itself, without the existence of a loop.}
\usage{
\method{reachable}{tempnetwork}(tempNetwork, startPoint, endPoint,
  safe = TRUE)
}
\arguments{
\item{tempNetwork}{A 'tempnetwork'-object.}

\item{startPoint}{The point-name (or point-id) of the start-point.}

\item{endPoint}{The point-name (or point-id) of the end-point.}

\item{safe}{If TRUE, it will be tested, whether the given point-names (or point-ids) are present in the 'tempflow'-object of the given 'tempnetwork'-object.}
}
\value{
Is TRUE, if there exists a path in 'tempflow'-object of the given 'tempnetwork'-object starting from "startPoint" and ending at "endPoint", else FALSE will be returned.
}
\description{
Is a function that takes a 'tempnetwork'-object and two point-names (or point-ids) and
tests whether one point ("endPoint") can be reached from the other ("startPoint") in the 'tempflow'-object of the given 'tempnetwork'-object.
A point can not reach itself, without the existence of a loop.
}
