% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tempnetwork-basic.R
\name{remove_gedges.tempnetwork}
\alias{remove_gedges.tempnetwork}
\title{This function allows one to remove edges from the specified graphs contained in the graph-list of the given 'tempnetwork'-object.
The vectors "pointNameList", "startTVertexList" and "endTVertexList" must all be of equal length, as their entries correspond with each other
That is, the i-th entry from the vector "pointNameList" specifies the graph from which the i-th edge ought to be removed, while "startTVertexList" and "endTVertexList" characterise
the start- and end-vertex of the i-th edge.}
\usage{
\method{remove_gedges}{tempnetwork}(tempNetwork, pointNameList,
  startTVertexList, endTVertexList, safe = TRUE)
}
\arguments{
\item{tempNetwork}{A 'tempnetwork'-object.}

\item{pointNameList}{A vector of strings representing the names of points in the 'tempflow'-object of the given 'tempnetwork'-object.}

\item{startTVertexList}{A vector of strings representing the names equivalence-classes in the equivalence-relation of the given 'tempnetwork'-object. Using the corresponding
entries in the "pointNameList"-vector, said equivalence-relation will be queried to determine the start-vertices of the edges to be removed}

\item{endTVertexList}{A vector of strings representing the names equivalence-classes in the equivalence-relation of the given 'tempnetwork'-object. Using the corresponding
entries in the "pointNameList"-vector, said equivalence-realtion will be queried to determine the end-vertices of the edges to be removed}

\item{safe}{If TRUE, it will be ensured that
the vectors "pointNameList", "startTVertexList" and "endTVertexList" are of equal length;
the specified edges actually exist in the corresponding graph;}
}
\value{
A new 'tempnetwork'-object, containing the same 'tempflow'-object as provided through the given 'tempnetwork'-object;
a new graph-list constructed by removing the specified edges from the corresponding graphs in the graph-list of the given 'tempnetwork'-object;
the same equivalence-relation as provided through the given 'tempnetwork'-object.
Additionally, the "tempGraph"-field will be set to NULL.
}
\description{
This function allows one to remove edges from the specified graphs contained in the graph-list of the given 'tempnetwork'-object.
The vectors "pointNameList", "startTVertexList" and "endTVertexList" must all be of equal length, as their entries correspond with each other
That is, the i-th entry from the vector "pointNameList" specifies the graph from which the i-th edge ought to be removed, while "startTVertexList" and "endTVertexList" characterise
the start- and end-vertex of the i-th edge.
}
