% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tempflow-basics.R
\name{remove_point.tempflow}
\alias{remove_point.tempflow}
\title{This function takes a 'tempflow'-object and a point-name (or point-id), which will be removed from said flow.
If the point has only a successor or only a predecessor, then only the incident edges will be removed. Otherwise,
all successors and all predecessors will be connected in a manner such that distances, i.e. weights are preserved.}
\usage{
\method{remove_point}{tempflow}(tempFlow, removePoint, safe = TRUE)
}
\arguments{
\item{tempFlow}{A 'tempflow'-object.}

\item{removePoint}{A point-name (or point-id) representing the point in time to be removed.}

\item{safe}{If TRUE, it will be ensured, that the given point is present in the specified 'tempflow'-object.
Additionally, the usual tempflow condition are checked (e.g. duplicate names, cycles, ... ).}
}
\value{
A new 'tempflow'-object, by removing the specified point and (if necessary) reconnecting the given 'tempflow'-object.
}
\description{
This function takes a 'tempflow'-object and a point-name (or point-id), which will be removed from said flow.
If the point has only a successor or only a predecessor, then only the incident edges will be removed. Otherwise,
all successors and all predecessors will be connected in a manner such that distances, i.e. weights are preserved.
}
