% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tempnetwork-basic.R
\name{remove_tvertices.tempnetwork}
\alias{remove_tvertices.tempnetwork}
\title{This function takes a list of equivalence-class names (i.e. tempvertices) and creates a new 'tempnetwork'-object by removing the
specified equivalence-classes from the equivalence-relation of the given 'tempnetwork'-object; removing all vertices contained in
in the specified equivalence-classes from their respective graphs in the graph-list of the given 'tempnetwork'-object.}
\usage{
\method{remove_tvertices}{tempnetwork}(tempNetwork, tVertexNameList,
  safe = TRUE)
}
\arguments{
\item{tempNetwork}{A 'tempnetwork'-object.}

\item{tVertexNameList}{A vector of strings representing the names of equivalence-classes in the equivalence-relation of the given 'tempnetwork'-object.}

\item{safe}{If TRUE, it will be ensured that
the names in the vector "tVertexNameList" are actually names of equivalence-classes in the equivalence-relation of the given 'tempnetwork'-object;
the vector "tVertexNameList" does not contain duplicates.
the resulting 'tempnetwork'-object satisfies all tempnetwork conditions.}
}
\value{
A new 'tempnetwork'-object, containing
the same 'tempflow'-object as the one given through the provided 'tempnetwork'-object;
a graph-list constructed by removing the vertices contained in the given equivalence-classes from their respective graph in the graph-list of the given 'tempnetwork'-object.
Additionally, the "tempGraph"-field will be set to NULL.
}
\description{
This function takes a list of equivalence-class names (i.e. tempvertices) and creates a new 'tempnetwork'-object by removing the
specified equivalence-classes from the equivalence-relation of the given 'tempnetwork'-object; removing all vertices contained in
in the specified equivalence-classes from their respective graphs in the graph-list of the given 'tempnetwork'-object.
}
