% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tempnetwork-setter.R
\name{set_basic_properties.tempnetwork}
\alias{set_basic_properties.tempnetwork}
\title{This function behaves as follows.
If "tempFlow" is not NULL, this function replaces the existing 'tempflow'-object in the field "tempFlow" of the given 'tempnetwork'-object
by the given 'tempflow'-object.
If "graphList" is not NULL, this function replaces the existing graph-list in the field "graphList" of the given 'tempnetwork'-object
by the given graph-list.
If "equivRel" is not NULL, this function replaces the existing equivalence-relation in the field "equivRel" of the given 'tempnetwork'-object
by the given equivalence-relation.}
\usage{
\method{set_basic_properties}{tempnetwork}(tempNetwork, tempFlow = NULL,
  graphList = NULL, equivRel = NULL, setDefaultGraphNames = TRUE,
  safe = TRUE)
}
\arguments{
\item{tempNetwork}{A 'tempnetwork'-object.}

\item{tempFlow}{A 'tempflow'-object.}

\item{graphList}{A named list of "igraph"-objects.}

\item{equivRel}{A named list of named lists,
where the outer-level list length must be identical to the length of each vertex set in the graph list.
where the inner-level names must be identical with the names of the newly added points;
where the values must be identical to the vertex-names of the corresponding newly added graph.}

\item{setDefaultGraphNames}{If TRUE any graph containing a vertex without a name, will be replaced by a graph where the name of each
vertex is a string corresponding to its index in the respective 'igraph'-object. Moreover, any graph containing a vertex with a name that is not a string, will be
coerced into a string.}

\item{safe}{If TRUE, the usual 'tempnetwork' conditions will be checked.}
}
\value{
The given 'tempnetwork'-object where
if "tempFlow" is not NULL, the existing 'tempflow'-object in the field "tempFlow" of the given 'tempnetwork'-object
is replaced by the given 'tempflow'-object.
if "graphList" is not NULL, the existing graph-list in the field "graphList" of the given 'tempnetwork'-object
is replaced by the given graph-list.
if "equivRel" is not NULL, the existing equivalence-relation in the field "equivRel" of the given 'tempnetwork'-object
is replaced by the given equivalence-relation.
Additionally, the "tempGraph"-field will be set to NULL.
}
\description{
This function behaves as follows.
If "tempFlow" is not NULL, this function replaces the existing 'tempflow'-object in the field "tempFlow" of the given 'tempnetwork'-object
by the given 'tempflow'-object.
If "graphList" is not NULL, this function replaces the existing graph-list in the field "graphList" of the given 'tempnetwork'-object
by the given graph-list.
If "equivRel" is not NULL, this function replaces the existing equivalence-relation in the field "equivRel" of the given 'tempnetwork'-object
by the given equivalence-relation.
}
