% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tempnetwork-setter.R
\name{set_equivRelation.tempnetwork}
\alias{set_equivRelation.tempnetwork}
\title{This function replaces the existing equivalence-relation in the field "equivRel" of the given 'tempnetwork'-object
by the given equivalence-relation.}
\usage{
\method{set_equivRelation}{tempnetwork}(tempNetwork, equivRel,
  safe = TRUE)
}
\arguments{
\item{tempNetwork}{A 'tempnetwork'-object.}

\item{equivRel}{A named list of named lists,
where the outer-level list length must be identical to the length of each vertex set in the graph list.
where the inner-level names must be identical with the names of the newly added points;
where the values must be identical to the vertex-names of the correspondign newly added graph.}

\item{safe}{If TRUE, the usual 'tempnetwork' conditions will be checked.}
}
\value{
The given 'tempnetwork'-object where the existing equivalence-relation in the field "equivRel" of the given 'tempnetwork'-object
is replaced by the given equivalence-relation. If the parameter "equivRel" is NULL the given unchanged 'tempnetwork'-object will be returned.
Additionally, the "tempGraph"-field will be set to NULL.
}
\description{
This function replaces the existing equivalence-relation in the field "equivRel" of the given 'tempnetwork'-object
by the given equivalence-relation.
}
