% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tempnetwork-setter.R
\name{set_graphList.tempnetwork}
\alias{set_graphList.tempnetwork}
\title{This function replaces the existing graph-list in the field "graphList" of the given 'tempnetwork'-object
by the given graph-list.}
\usage{
\method{set_graphList}{tempnetwork}(tempNetwork, graphList,
  setDefaultGraphNames = TRUE, safe = TRUE)
}
\arguments{
\item{tempNetwork}{A 'tempnetwork'-object.}

\item{graphList}{A named list of "igraph"-objects.}

\item{setDefaultGraphNames}{If TRUE any graph containing a vertex without a name, will be replaced by a graph where the name of each
vertex is a string corresponding to its index in the respective 'igraph'-object. Moreover, any graph containing a vertex with a name that is not a string, will be
coerced into a string.}

\item{safe}{If TRUE, the usual 'tempnetwork' conditions will be checked.}
}
\value{
The given 'tempnetwork'-object where the existing graph-list in the field "graphList" of the given 'tempnetwork'-object
is replaced by the given graph-list. If the parameter "graphList" is NULL the given unchanged 'tempnetwork'-object will be returned.
Additionally, the "tempGraph"-field will be set to NULL.
}
\description{
This function replaces the existing graph-list in the field "graphList" of the given 'tempnetwork'-object
by the given graph-list.
}
