% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tempflow-setter.R
\name{set_point_names.tempflow}
\alias{set_point_names.tempflow}
\title{Given a vector containing the current point-names (or point-ids) and a vector containing strings,
this function changes the names of the specified points to the strings in the 'nameList'-vector.}
\usage{
\method{set_point_names}{tempflow}(tempFlow, pointList = tP(tempFlow),
  nameList = NULL, safe = TRUE)
}
\arguments{
\item{tempFlow}{A 'tempflow'-object.}

\item{pointList}{A vector containing the point-names (or point-ids) of the points which names ought to be changed.}

\item{nameList}{A vector containing strings that ought to become the new names of the specified points.}

\item{safe}{If TRUE, it will be ensured that the specified point-names (or point-ids) are present in the given 'tempflow'-object;
the 'pointList'-vector and 'nameList'-vector are of the same length;
that point-names remain unique;
that point-names remain strings.}
}
\value{
A 'tempflow'-object, where the point-names of the points specified in the 'pointList'-vector are
assigned the names given by the 'nameList'-vector.
}
\description{
Given a vector containing the current point-names (or point-ids) and a vector containing strings,
this function changes the names of the specified points to the strings in the 'nameList'-vector.
}
