% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tempnetwork-tempflow-setter.R
\name{set_point_names.tempnetwork}
\alias{set_point_names.tempnetwork}
\title{Given a vector containing the current point-names (or point-ids) and a vector containing strings,
this function changes the names of the specified points in the 'tempflow'-object of the given 'tempnetwork'-object to the strings in the 'nameList'-vector.
Moreover, the corresponding names of the equivalence relation will be changed as well.}
\usage{
\method{set_point_names}{tempnetwork}(tempNetwork,
  pointList = tP(tempNetwork), nameList = NULL, safe = TRUE)
}
\arguments{
\item{tempNetwork}{A 'tempnetwork'-object.}

\item{pointList}{A vector containing the point-names (or point-ids) of the points which names ought to be changed.}

\item{nameList}{A vector containing strings that ought to become the new names of the specified points.}

\item{safe}{If TRUE, it will be ensured that the specified point-names (or point-ids) are present in the 'tempflow'-object of the given 'tempnetwork'-object;
the 'pointList'-vector and 'nameList'-vector are of the same length;
that point-names remain unique;
that point-names remain strings.}
}
\value{
A 'tempnetwork'-object, with a 'tempflow'-object where the point-names of the points specified in the 'pointList'-vector are
assigned the names given by the 'nameList'-vector; with a equivalence relation where the point-names of the points specified in the 'pointList'-vector are
assigned the names given by the 'nameList'-vector.
}
\description{
Given a vector containing the current point-names (or point-ids) and a vector containing strings,
this function changes the names of the specified points in the 'tempflow'-object of the given 'tempnetwork'-object to the strings in the 'nameList'-vector.
Moreover, the corresponding names of the equivalence relation will be changed as well.
}
