% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tempnetwork-tempflow-setter.R
\name{set_step_weights_from_id.tempnetwork}
\alias{set_step_weights_from_id.tempnetwork}
\title{Given a vector containing the current step-ids and a vector containing numeric values (excluding NA, NaN and NULL),
this function sets the weights of the specified steps  in the 'tempflow'-object (of the given 'tempnetwork'-object) to the values given by the "weightList"-vector.}
\usage{
\method{set_step_weights_from_id}{tempnetwork}(tempNetwork,
  stepIDList = NULL, weightList = NULL, safe = TRUE)
}
\arguments{
\item{tempNetwork}{A 'tempnetwork'-object.}

\item{stepIDList}{A vector containing the step-ids of the steps which weights ought to be changed.}

\item{weightList}{Is a vector containing numeric values (excluding NA, NaN and NULL) representing the step-weights of the given steps
If left NULL, every newly introduced loop will be assigned the value 0, and every other new edge will be assigned the value 1.
Otherwise, it must be either a scalar (i.e. length 1) or a vector with the same length as the "stepIDList"-vector.}

\item{safe}{If TRUE, it will be ensured that the specified step-ids exist;
that the "weightList"-vector' is either of length 1 or has the same length as the "stepIDList"-vector;
that the weights of the resulting 'tempflow'-object are non-negative, numerical values.}
}
\value{
A 'tempnetwork'-object, with a 'tempflow'-object where the step-weights of the steps specified in the "stepIDList"-vector are
assigned the weights given by the "weightList"-vector.
}
\description{
Given a vector containing the current step-ids and a vector containing numeric values (excluding NA, NaN and NULL),
this function sets the weights of the specified steps  in the 'tempflow'-object (of the given 'tempnetwork'-object) to the values given by the "weightList"-vector.
}
