% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tempnetwork-setter.R
\name{set_tempFlow.tempnetwork}
\alias{set_tempFlow.tempnetwork}
\title{This function replaces the existing 'tempflow'-object in the field "tempFlow" of the given 'tempnetwork'-object
by the given 'tempflow'-object.}
\usage{
\method{set_tempFlow}{tempnetwork}(tempNetwork, tempFlow, safe = TRUE)
}
\arguments{
\item{tempNetwork}{A 'tempnetwork'-object.}

\item{tempFlow}{A 'tempflow'-object.}

\item{safe}{If TRUE, the usual 'tempnetwork' conditions will be checked.}
}
\value{
The given 'tempnetwork'-object where the existing 'tempflow'-object in the field "tempFlow" of the given 'tempnetwork'-object
is replaced by the given 'tempflow'-object. If the parameter "tempFlow" is NULL the given unchanged 'tempnetwork'-object will be returned.
Additionally, the "tempGraph"-field will be set to NULL.
}
\description{
This function replaces the existing 'tempflow'-object in the field "tempFlow" of the given 'tempnetwork'-object
by the given 'tempflow'-object.
}
