% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tempflow-basics.R
\name{slice_flow.tempflow}
\alias{slice_flow.tempflow}
\title{Is a function that takes a 'tempflow'-object and a vector of start-points in time and a vector of end-points in time.
Given those, the cross-product of "startPointList" and "endPointList" will be computed, from there any point that lies
between the points of any tuple will be included in the new 'tempflow'-object.
Given those, a new flow will be returned that contains all points (including start point and end point).}
\usage{
\method{slice_flow}{tempflow}(tempFlow, startPointList, endPointList,
  safe = TRUE)
}
\arguments{
\item{tempFlow}{A 'tempflow'-object.}

\item{startPointList}{A vector containing point-names (or point-ids) representing the names of the start-points of the new 'tempflow'-object.}

\item{endPointList}{A vector containing point-names (or point-ids) representing the names of the end-points of the new 'tempflow'-object.}

\item{safe}{If TRUE, it will be ensured that all the specified points exist in the 'tempflow'-object and
that all relevant 'tempflow'-conditions are uphold.}
}
\value{
A new 'tempflow'-object, reduced to the points in time between the "startPoint" and the "endPoint".
}
\description{
Is a function that takes a 'tempflow'-object and a vector of start-points in time and a vector of end-points in time.
Given those, the cross-product of "startPointList" and "endPointList" will be computed, from there any point that lies
between the points of any tuple will be included in the new 'tempflow'-object.
Given those, a new flow will be returned that contains all points (including start point and end point).
}
