% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tempnetwork-tempflow-basics.R
\name{slice_flow.tempnetwork}
\alias{slice_flow.tempnetwork}
\title{Is a function that takes a 'tempnetwork'-object and a vector of start-points in time and a vector of end-points in time.
Given those, the cross-product of "startPointList" and "endPointList" will be computed, from there any point that lies
between the points of any tuple will be included in the new 'tempflow'-object contained in the resulting 'tempnetwork'-object.
Given those, a new flow will be returned that contains all points (including start point and end point).}
\usage{
\method{slice_flow}{tempnetwork}(tempNetwork, startPointList, endPointList,
  safe = TRUE)
}
\arguments{
\item{tempNetwork}{A 'tempnetwork'-object.}

\item{startPointList}{A vector containing point-names (or point-ids) representing the names of the start-points of the new 'tempflow'-object.}

\item{endPointList}{A vector containing point-names (or point-ids) representing the names of the end-points of the new 'tempflow'-object.}

\item{safe}{If TRUE, it will be ensured that all the specified points exist in the 'tempflow'-object and
that all relevant 'tempflow' and 'tempnetwork'-conditions are uphold.}
}
\value{
A new 'tempnetwork'-object, containing a new 'tempflow'-object reduced to the points in time between the "startPoint" and the "endPoint".
 containing a new "graphList" constructed by restricting the existing "graphList" to those entries corresponding to the points in the  new 'tempflow'-object;
 containing a new "equivRel" constructed by restricting the existing "equivRel" to those entries corresponding to the points in the  new 'tempflow'-object.
}
\description{
Is a function that takes a 'tempnetwork'-object and a vector of start-points in time and a vector of end-points in time.
Given those, the cross-product of "startPointList" and "endPointList" will be computed, from there any point that lies
between the points of any tuple will be included in the new 'tempflow'-object contained in the resulting 'tempnetwork'-object.
Given those, a new flow will be returned that contains all points (including start point and end point).
}
