% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tempnetwork-getter.R
\name{stG.tempnetwork}
\alias{stG.tempnetwork}
\title{Returns the flag indicating whether the graph representation of the 'tempnetwork'-object
should be stored.}
\usage{
\method{stG}{tempnetwork}(tempNetwork)
}
\arguments{
\item{tempNetwork}{A 'tempnetwork'-object.}
}
\value{
A boolean value indicating whether the graph representation of the 'tempnetwork'-object should be stored.
}
\description{
Returns the flag indicating whether the graph representation of the 'tempnetwork'-object
should be stored.
}
