% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tempflow.R
\name{tempflow}
\alias{tempflow}
\title{Creates flow of time. Two major methods of addressing points within the 'tempflow'-object, are either by name or by id.
Names have to be string, meaning that if addressed by name it must be passed as string.
Similarly, if addressed by id, the value passed must be an integer.}
\usage{
tempflow(inpGraph, setDefaultValues = FALSE, safe = TRUE)
}
\arguments{
\item{setDefaultValues}{If TRUE, any invalid weight or name assignment will be replaced by its corresponding default assignments.
\itemize{
  \item Under the default weight assignment, the weight of a loop-edge is 0 and of a non-loop-edge is 1.
  \item Under the default name assignment, the name of a vertex is its current vertex-id (as assigned within the igraph package) as a string.
}}

\item{safe}{If TRUE, it will be tested, whether inpGraph
\itemize{
  \item is an 'igraph'-object;
  \item is a directed and acyclic;
  \item contains no duplicate edges;
  \item has unique vertex names (of type string);
  \item whether the edge attribute "weight" takes on non-negative, numerical values for each edge in the graph.
}}

\item{G}{A directed acyclic 'igraph'-object.}
}
\value{
A flow of time object.
}
\description{
Creates flow of time. Two major methods of addressing points within the 'tempflow'-object, are either by name or by id.
Names have to be string, meaning that if addressed by name it must be passed as string.
Similarly, if addressed by id, the value passed must be an integer.
}
