% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tempnetwork-getter.R
\name{tF.tempnetwork}
\alias{tF.tempnetwork}
\title{Returns the 'tempflow'-object of the given 'tempnetwork'-object.}
\usage{
\method{tF}{tempnetwork}(tempNetwork)
}
\arguments{
\item{tempNetwork}{A 'tempnetwork'-object.}
}
\value{
The 'tempflow'-object of the given 'tempnetwork'-object.
}
\description{
Returns the 'tempflow'-object of the given 'tempnetwork'-object.
}
