% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tempnetwork-getter.R
\name{tG.tempnetwork}
\alias{tG.tempnetwork}
\title{Returns the value stored in the field "tempGraph" of the given 'tempnetwork'-object. This can either be the value NULL
or the graph representation of the given 'tempnetwork'-object.}
\usage{
\method{tG}{tempnetwork}(tempNetwork, computeTempGraph = FALSE,
  connectGraphs = TRUE)
}
\arguments{
\item{tempNetwork}{A 'tempnetwork'-object.}

\item{computeTempGraph}{If TRUE the graph representation of the given 'tempnetwork'-object will be computed (if necessary).}
}
\value{
The value stored in the field "tempGraph" of the given 'tempnetwork'-object.
}
\description{
Returns the value stored in the field "tempGraph" of the given 'tempnetwork'-object. This can either be the value NULL
or the graph representation of the given 'tempnetwork'-object.
}
