% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tempnetwork-getter.R
\name{tGvN.tempnetwork}
\alias{tGvN.tempnetwork}
\title{Using the tVertex name and the point name this function computes the name of the corresponding vertex
in the tempGraph representation of the given 'tempnetwork'-object.}
\usage{
\method{tGvN}{tempnetwork}(tempNetwork, tVertexName, pointName,
  safe = TRUE)
}
\arguments{
\item{tempNetwork}{A 'tempnetwork'-object.}

\item{tVertexName}{A string corresponding with the name of an equivalence class specified in the "equigRel"  of the given 'tempnetwork'-object.}

\item{pointName}{A string corresponding with the name of a point in the 'tempflow'-object of the given 'tempnetwork'-object.}

\item{safe}{If TRUE, it will be ensured that "tVertexName" and "pointName" exist.}
}
\value{
A string representing the name of the corresponding vertex in the tempGraph representation of the given 'tempnetwork'-object.
}
\description{
Using the tVertex name and the point name this function computes the name of the corresponding vertex
in the tempGraph representation of the given 'tempnetwork'-object.
}
