% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tempflow-getter.R
\name{tP.tempflow}
\alias{tP.tempflow}
\title{Returns the points in time of the given 'tempflow'-object.}
\usage{
\method{tP}{tempflow}(tempFlow, ...)
}
\arguments{
\item{tempFlow}{A 'tempflow'-object.}
}
\value{
A list of 'igraph'-vertices (from the underlying 'igraph'-object),
representing the points in time of the 'tempflow'-object.
}
\description{
Returns the points in time of the given 'tempflow'-object.
}
