% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tempnetwork-tempflow-getter.R
\name{tP.tempnetwork}
\alias{tP.tempnetwork}
\title{Returns the points in time from the 'tempflow'-object of the given 'tempnetwork'-object.}
\usage{
\method{tP}{tempnetwork}(tempNetwork, ...)
}
\arguments{
\item{tempNetwork}{A 'tempnetwork'-object.}
}
\value{
A list of 'igraph'-vertices (from the underlying 'igraph'-object),
representing the points in time from the 'tempflow'-object of the given 'tempnetwork'-object.
}
\description{
Returns the points in time from the 'tempflow'-object of the given 'tempnetwork'-object.
}
