% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tempflow-getter.R
\name{tS.tempflow}
\alias{tS.tempflow}
\title{Returns the steps in time of the given 'tempflow'-object.}
\usage{
\method{tS}{tempflow}(tempFlow, ...)
}
\arguments{
\item{tempFlow}{A 'tempflow'-object.}
}
\value{
A list of 'igraph'-edges (from the underlying 'igraph'-object),
representing the steps in time of the 'tempflow'-object.
}
\description{
Returns the steps in time of the given 'tempflow'-object.
}
