% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tempnetwork-tempflow-getter.R
\name{tS.tempnetwork}
\alias{tS.tempnetwork}
\title{Returns the steps in time from the 'tempflow'-object of the given 'tempnetwork'-object.}
\usage{
\method{tS}{tempnetwork}(tempNetwork, ...)
}
\arguments{
\item{tempNetwork}{A 'tempnetwork'-object.}
}
\value{
A list of 'igraph'-edges (from the underlying 'igraph'-object),
representing the steps in time from the 'tempflow'-object of the given 'tempnetwork'-object.
}
\description{
Returns the steps in time from the 'tempflow'-object of the given 'tempnetwork'-object.
}
