% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tempnetwork-getter.R
\name{tV.tempnetwork}
\alias{tV.tempnetwork}
\title{Returns the names of the temporal vertices (i.e. the names of the equivalence classes in the "equivRel") of the given 'tempnetwork'-object}
\usage{
\method{tV}{tempnetwork}(tempNetwork)
}
\arguments{
\item{tempNetwork}{A 'tempnetwork'-object.}
}
\value{
The names of the temporal vertices (i.e. the names of the equivalence classes in the "equivRel") of the given 'tempnetwork'-object
}
\description{
Returns the names of the temporal vertices (i.e. the names of the equivalence classes in the "equivRel") of the given 'tempnetwork'-object
}
